% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/during-functions.R
\name{first_during_after_event}
\alias{first_during_after_event}
\alias{nearest_to_event_during}
\title{During, adjacent to, functions}
\usage{
first_during_after_event(..., event_datetime, names_suffix = "")

nearest_to_event_during(..., event_datetime, names_suffix = "")
}
\description{
During, adjacent to, functions
}
