% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm.R
\name{adm_rename}
\alias{adm_rename}
\title{Tidy raw tests colnames}
\usage{
adm_rename(x, names = default_rename("adm"))
}
\arguments{
\item{x}{A data frame of raw adm data}

\item{names}{A vector of new_name = old_name pairs}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument
}
\author{
R.J.B. Goudie
}
