% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm.R
\name{adm_unrename}
\alias{adm_unrename}
\title{Untidy raw adm colnames}
\usage{
adm_unrename(x)
}
\arguments{
\item{x}{A data frame of raw adm data with tidy names}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
Untidy raw adm colnames
}
\author{
R.J.B. Goudie
}
