% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt.R
\name{adt_annotate}
\alias{adt_annotate}
\title{Annotate ADT data}
\usage{
adt_annotate(x, fixed_labels, annotate_fn)
}
\arguments{
\item{x}{Tidy ADT data, as tidied by \code{\link[=adt_rename]{adt_rename()}}}

\item{fixed_labels}{A data frame with a colum \code{department}, which will be
joined to the ADT data}

\item{annotate_fn}{A function that provides additional annotation}
}
\value{
The supplied tidy ADT data \code{x} with additional annotations
}
\description{
Removes duplicate rows
}
\details{
Preforms the following checks
\itemize{
\item Checks discharge dates are identical within a visit
}

Annotates the ADT data with
\itemize{
\item \code{visit_index}
\item \code{end_datetime}
}
}
\examples{
fixed_labels <-
  tibble::tribble(
    ~department, ~is_ed,
    "ADD EMERGENCY DEPT", TRUE,
    "ADD C2 WARD", FALSE,
    "ADD D2 WARD", FALSE,
    "DIALYSIS CENTRE", FALSE,
    "ADD A1 WARD", FALSE,
    "ADD A2 WARD", FALSE,
    "ADD A3 WARD", FALSE,
    "ADD A4 WARD", FALSE,
    "ADD X WARD", FALSE,
    "ADD Y WARD", FALSE,
    "ADD Z WARD", FALSE)

adt_data_example <-
  read_tidybrookes_csv(
    file = tidybrookes_example("adt.csv"),
    col_types = "adt"
  ) \%>\%
  adt_rename

adt_data_example \%>\%
  adt_annotate(annotate_fn = identity,
               fixed_labels = fixed_labels)

}
\author{
R.J.B. Goudie
}
