% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt.R
\name{adt_check_discharge_dates_identical}
\alias{adt_check_discharge_dates_identical}
\title{Check discharge dates identical within a visit}
\usage{
adt_check_discharge_dates_identical(x)
}
\description{
All discharge_datetime should be the same within a visit
}
