% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt.R
\name{adt_department_summary}
\alias{adt_department_summary}
\title{Summarise department visits data in ADT data}
\usage{
adt_department_summary(x)
}
\arguments{
\item{x}{Tidy ADT data, as tidied by \code{\link[=adt_rename]{adt_rename()}}}
}
\value{
The supplied tidy ADT data \code{x}, with an additional columns
}
\description{
Annotates department visits with \code{department_visit_index}, which indexes
the department visits WITHIN a \code{visit_id}.
}
\details{
Also adds \code{department_start_datetime}, \code{department_end_datetime},
\code{department_length_days}.
}
\author{
R.J.B. Goudie
}
