% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt.R
\name{adt_rename}
\alias{adt_rename}
\title{Tidy raw tests colnames}
\usage{
adt_rename(x)
}
\arguments{
\item{x}{A data frame of raw adt data}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
Tidy raw tests colnames
}
\author{
R.J.B. Goudie
}
