% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny0.R
\name{check_data}
\alias{check_data}
\title{Tidy raw tests colnames}
\usage{
check_data(
  adm_data,
  adt_data,
  tests_data,
  fsheet_data,
  diagnosis_pl_raw,
  condition_data
)
}
\arguments{
\item{x}{A data frame of raw test data}

\item{names}{A vector of new_name = old_name pairs}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument
}
\author{
R.J.B. Goudie
}
