% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_that_all}
\alias{check_that_all}
\title{Check all rows of a data frame satisfy a condition}
\usage{
check_that_all(x, condition, name = "Unnamed", summary = identity)
}
\arguments{
\item{x}{A data frame}

\item{condition}{A condition}

\item{name}{Character string, a name for the condition}
}
\description{
Check all rows of a data frame satisfy a condition
}
