% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{condition_add}
\alias{condition_add}
\title{Add a new definition of a Diagnosis PL item}
\usage{
condition_add(
  condition_def,
  symbol,
  title,
  icd10 = NA,
  condition_start_datetime_fn = case_when(!is.na(diagnosis_datetime) ~
    diagnosis_datetime, !is.na(entered_datetime) ~ entered_datetime, TRUE ~ ymd_hms(NA)),
  condition_end_datetime_fn = case_when(!is.na(resolved_datetime) ~ resolved_datetime,
    TRUE ~ ymd_hms(NA)),
  silently_exclude_deleted_when = TRUE,
  silently_exclude_resolved_when = FALSE
)
}
\arguments{
\item{condition_def}{A list of existing condition definitions to add
this to}

\item{symbol}{A short name for the condition item. This should be a
character vector that is suitable to use as an R object name}

\item{title}{The full name for the condition item, suitable for using in
figure titles etc}

\item{icd_10}{A vector of ICD-10 codes to use.}

\item{silently_exclude_deleted}{When TRUE, diagnoses with
\code{status = "Deleted"} will be removed from the data frame. In theory, only
diagnoses made in error should be deleted, but in practice it is thought
that clinical staff often delete diagnoses that are resolved rather than
mark them as "Resolved".}
}
\description{
Add a new definition of a Diagnosis PL item
}
\author{
R.J.B. Goudie
}
