% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{condition_extract}
\alias{condition_extract}
\title{Extract condition data into tidy format}
\usage{
condition_extract(x, condition_def, errors = stop)
}
\arguments{
\item{x}{condition data in tidy-column-name format (after applying
\code{condition_rename})}

\item{condition_def}{A condition definition}

\item{errors}{A function indicating what to do when an error is found}
}
\value{
A data frame with the following columns:
\code{person_id}, \code{symbol}, \code{value_as_number}, \code{value_as_character},
\code{censoring}, \code{ordered_datetime}, \code{collected_datetime}, \code{received_datetime},
\code{result_datetime}, \code{ordering_department_name}, \code{range_low}, \code{range_high},
\code{unit}, \code{name}, \code{title}, \code{method}, \code{group}, \code{type}

The result will be sorted by diagnosis_datetime
}
\description{
Extract condition data into tidy format
}
\author{
R.J.B. Goudie
}
