% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogs.R
\name{demogs_unrename}
\alias{demogs_unrename}
\title{Untidy raw demogs colnames}
\usage{
demogs_unrename(
  x,
  names = c(STUDY_SUBJECT_DIGEST = "person_id", GENDER_DESC = "gender")
)
}
\arguments{
\item{x}{A data frame of raw demogs data with tidy names}

\item{names}{A vector of new_name = old_name pairs}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument
}
\author{
R.J.B. Goudie
}
