% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_pl.R
\name{diagnosis_pl_add}
\alias{diagnosis_pl_add}
\title{Add a new definition of a Diagnosis PL item}
\usage{
diagnosis_pl_add(
  diagnosis_pl_def,
  symbol,
  title,
  icd10 = NA,
  silently_exclude_deleted = TRUE
)
}
\arguments{
\item{diagnosis_pl_def}{A list of existing diagnosis_pl definitions to add
this to}

\item{symbol}{A short name for the diagnosis_pl item. This should be a
character vector that is suitable to use as an R object name}

\item{title}{The full name for the diagnosis_pl item, suitable for using in
figure titles etc}

\item{silently_exclude_deleted}{When TRUE, diagnoses with
\code{status = "Deleted"} will be removed from the data frame. In theory, only
diagnoses made in error should be deleted, but in practice it is thought
that clinical staff often delete diagnoses that are resolved rather than
mark them as "Resolved".}

\item{icd_10}{A vector of ICD-10 codes to use.}
}
\description{
Add a new definition of a Diagnosis PL item
}
\author{
R.J.B. Goudie
}
