% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_pl.R
\name{diagnosis_pl_extract}
\alias{diagnosis_pl_extract}
\title{Extract diagnosis_pl data into tidy format}
\usage{
diagnosis_pl_extract(x, diagnosis_pl_def, errors = stop)
}
\arguments{
\item{x}{Diagnosis_pl data in tidy-column-name format (after applying
\code{diagnosis_pl_rename})}

\item{diagnosis_pl_def}{A diagnosis_pl definition}

\item{errors}{A function indicating what to do when an error is found}
}
\value{
A data frame with the following columns:
\code{person_id}, \code{symbol}, \code{value_as_number}, \code{value_as_character},
\code{censoring}, \code{ordered_datetime}, \code{collected_datetime}, \code{received_datetime},
\code{result_datetime}, \code{ordering_department_name}, \code{range_low}, \code{range_high},
\code{unit}, \code{name}, \code{title}, \code{method}, \code{group}, \code{type}

The result will be sorted by diagnosis_date
}
\description{
Extract diagnosis_pl data into tidy format
}
\author{
R.J.B. Goudie
}
