% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_pl.R
\name{diagnosis_pl_rename}
\alias{diagnosis_pl_rename}
\title{Tidy raw diagnosis_pl colnames}
\usage{
diagnosis_pl_rename(x)
}
\arguments{
\item{x}{A data frame of raw diagnosis_pl data}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
Tidy raw diagnosis_pl colnames
}
\author{
R.J.B. Goudie
}
