% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{distinct_inform}
\alias{distinct_inform}
\title{Distinct, but report the filtering that occurs}
\usage{
distinct_inform(x)
}
\arguments{
\item{x}{A data frame}

\item{since}{A character describing the reason for the filtering}
}
\description{
Distinct, but report the filtering that occurs
}
