% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{duckdb_read_tidybrookes_csv}
\alias{duckdb_read_tidybrookes_csv}
\title{Load data from CUH in delimited (including CSV) format into duckdb}
\usage{
duckdb_read_tidybrookes_csv(
  src,
  table,
  file,
  col_types = table,
  na = c("", "NA"),
  quote = "\\"",
  table_name = paste0(table, "_raw_unrenamed"),
  strict_mode = TRUE,
  append = FALSE
)
}
\arguments{
\item{src}{A \code{DBIConnection} object produced by \link{duckdb_tidybrookes_connect}
or \code{DBI::dbConnect()}}

\item{table}{A string specifying the name of the table that the
data comes from, for which a standard format can be used. Available values
are \code{adm}, \code{adt}, \code{demogs}, \code{fsheet}, \code{fsheet_io}, \code{tests}, \code{med_admin},
\code{med_prescr}, \code{diagnosis_pl}, \code{med_hist}, \code{radiology}. This is translated
into column specification using \code{\link{default_col_types}}.}

\item{file}{A path to the CSV file}

\item{col_types}{Either a string specifying the name of the table that the
data comes from, for which a standard format can be used. Available values
are \code{adm}, \code{adt}, \code{demogs}, \code{fsheet}, \code{fsheet_io}, \code{tests}, \code{med_admin},
\code{med_prescr}, \code{diagnosis_pl}, \code{med_hist}, \code{radiology}. This is translated
into column specification using \code{\link{default_col_types}}.

Or a \code{\link[readr:cols]{readr::cols()}} format list of column names and types, which can be
used where a nonstandard data format are supplied.}

\item{na}{Character vector of strings to interpret as missing values. Set
this option to \code{character()} to indicate no missing values.}

\item{quote}{Single character used to quote strings}

\item{table_name}{The name to use for the table in \code{duckdb}}

\item{strict_mode}{Should \code{duckdb}'s strict mode for loading CSVs be used?}

\item{append}{Should new data be appended to the end of an existing table in
the database?}
}
\description{
Convenience wrapper that makes it easier to set the \code{col_types} for
standard data from CUH for loading into \code{duckdb}.
}
\author{
R.J.B. Goudie
}
