% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{duckdb_tidybrookes_connect}
\alias{duckdb_tidybrookes_connect}
\title{Connect to duckdb database with suitable defaults}
\usage{
duckdb_tidybrookes_connect(
  db_dir = "",
  bigint = "integer64",
  timezone_out = "Europe/London",
  tz_out_convert = "force",
  temp_dir = NULL
)
}
\arguments{
\item{db_dir}{Location for database files. Should be a path to an existing
directory in the file system. With the default (or \code{""}), all data is
kept in RAM.}

\item{timezone_out}{The time zone returned to R, defaults to \code{"UTC"}, which
is currently the only timezone supported by duckdb.
If you want to display datetime values in the local timezone,
set to \code{\link[=Sys.timezone]{Sys.timezone()}} or \code{""}.}

\item{tz_out_convert}{How to convert timestamp columns to the timezone
specified in \code{timezone_out}. There are two options: \code{"with"}, and
\code{"force"}. If \code{"with"} is chosen, the timestamp will be returned as it
would appear in the specified time zone.
If \code{"force"} is chosen, the timestamp will have the same clock
time as the timestamp in the database, but with the new time}

\item{temp_dir}{Set the directory to which duckdb should write temp files.
For data stored on slow network drives, it may be much faster to set
this to a faster local drive.}
}
\description{
This is a wrapper around \code{\link[DBI:dbConnect]{DBI::dbConnect()}} that sets suitable defaults
for use with \code{tidybrookes} datasets:
\itemize{
\item uses \link{bit64} \code{integer64} representation for big integers, since some
IDs from EHRs can be larger than can be represented by standard integers
and representing as floats/doubles is incorrect
\item Since duckdb support for timezone is not great in R currently, we store
datetimes in UTC in duckdb and force conversion to local time at export
from duckdb.
}
}
\details{
Facility to specify the \code{temp_directory} for duckdb is also provided
}
\examples{
\dontrun{
con <- duckdb_tidybrookes_connect()
}
}
\author{
R.J.B. Goudie
}
