% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/during-functions.R
\name{first_during}
\alias{first_during}
\alias{last_during}
\alias{nearest_visit_start_during}
\title{During functions}
\usage{
first_during(...)

last_during(...)

nearest_visit_start_during(...)
}
\description{
During functions
}
\examples{

# raw data that we will combine
adm_data_example
fsheet_news2_example

# extract the first of each fsheet symbol during the hospital visit
adm_data_example \%>\%
  first_during(fsheet_news2_example,
               datetime = measurement_datetime,
               during = "during_visit") \%>\%
  select(person_id,
         visit_id,
         visit_start_datetime,
         visit_end_datetime,
         contains("news2"))

# extract the last of each fsheet symbol during the hospital visit
adm_data_example \%>\%
 last_during(fsheet_news2_example,
             datetime = measurement_datetime,
             during = "during_visit") \%>\%
  select(person_id,
         visit_id,
         visit_start_datetime,
         visit_end_datetime,
         contains("news2"))

# can pipe together two summaries
adm_data_example \%>\%
  first_during(fsheet_news2_example,
               datetime = measurement_datetime,
               during = "during_visit") \%>\%
 last_during(fsheet_news2_example,
             datetime = measurement_datetime,
             during = "during_visit") \%>\%
  select(person_id,
         visit_id,
         visit_start_datetime,
         visit_end_datetime,
         contains("news2"))

# example data with 3 fsheet data fields
fsheet_example

# summaries for all symbols can be calculated together
adm_data_example \%>\%
  first_during(fsheet_example,
               datetime = measurement_datetime,
               during = "during_visit") \%>\%
  select(person_id,
         visit_id,
         visit_start_datetime,
         visit_end_datetime,
         contains("during"))
}
