% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{extract_col_types}
\alias{extract_col_types}
\title{List of standard column types for Epic data}
\usage{
extract_col_types(col_types)
}
\arguments{
\item{col_types}{Either a string specifying the name of the table that the
data comes from, for which a standard format can be used. Available
value are \code{adm}, \code{adt}, \code{demogs}, \code{fsheet}, \code{fsheet_io}, \code{tests},
\code{med_admin}, \code{med_prescr}, \code{diagnosis_pl}, \code{med_hist}, \code{radiology}.

Or a \code{\link[readr:cols]{readr::cols()}} format list of column names and types, which
can be used where a nonstandard data format are supplied.}
}
\description{
List of standard column types for Epic data
}
