% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_person}
\alias{filter_person}
\alias{filter_visit}
\alias{filter_tests}
\alias{filter_med_prescr}
\alias{filter_numeric}
\alias{filter_logical}
\alias{filter_character}
\title{Filtering functions}
\usage{
filter_person(x, ...)

filter_visit(x, visit_id)

filter_tests(x, pattern)

filter_med_prescr(x, pattern)

filter_numeric(x)

filter_logical(x)

filter_character(x)
}
\description{
Filtering functions
}
