% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_inform}
\alias{filter_inform}
\title{Filter, but report the filtering that occurs}
\usage{
filter_inform(x, ..., since = "for unknown reason")
}
\arguments{
\item{x}{A data frame}

\item{...}{Passed to filter}

\item{since}{A character describing the reason for the filtering}
}
\description{
Filter, but report the filtering that occurs
}
