% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fn_ignoring_missing}
\alias{fn_ignoring_missing}
\title{Apply function ignoring missing columns}
\usage{
fn_ignoring_missing(fn, x, ...)
}
\arguments{
\item{fn}{A function}

\item{x}{A data frame}

\item{...}{Arguments to \code{fn}, which should be columns names of \code{x}}
}
\description{
Apply function ignoring missing columns
}
