% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_as_argument}
\alias{format_as_argument}
\title{Convert a vector into a character vector, suitably formatted for use
in a R argument list - ie each element on a new line.}
\usage{
format_as_argument(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A character string of \code{x}
}
\description{
Convert a vector into a character vector, suitably formatted for use
in a R argument list - ie each element on a new line.
}
