% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsheet.R
\name{fsheet_add}
\alias{fsheet_add}
\title{Add a new definition of a flowsheet item}
\usage{
fsheet_add(
  fsheet_def,
  symbol,
  title,
  names = NA,
  search_pattern,
  search_exclude = NA,
  type = "numeric",
  silently_exclude_na_when = FALSE,
  silently_exclude_when = FALSE,
  censoring_fn = case_when(TRUE ~ NA_character_),
  value_as_number_fn = if_else(type == "numeric", case_when(TRUE ~ value_as_number),
    case_when(TRUE ~ NA_real_)),
  value_as_character_fn = if_else(type == "character", case_when(TRUE ~
    value_as_character), case_when(TRUE ~ NA_character_)),
  value_as_logical_fn = if_else(type == "logical", case_when(TRUE ~ value_as_logical),
    case_when(TRUE ~ as.logical(NA))),
  unit_rescale_fn = case_when(TRUE ~ value_as_number),
  unit_relabel_fn = case_when(TRUE ~ NA_character_),
  coalesce_fn = identity,
  expect_before = TRUE,
  expect_after = TRUE,
  range_mainly_low = NA,
  range_mainly_high = NA,
  range_discard_below = NA,
  range_discard_above = NA
)
}
\arguments{
\item{fsheet_def}{A list of existing fsheet definitions to add this to}

\item{symbol}{A short name for the flowsheet item. This should be a character vector that is suitable to use as an R object name}

\item{title}{The full name for the flowsheet item, suitable for using in
figure titles etc}

\item{names}{The display names (\code{disp_name} in raw format) of the flowsheet
items to include}

\item{search_pattern}{A charater vector of potential synonymns for this
flowsheet item, in case new names (\code{disp_name} in raw format) for the
same item are added in future.}

\item{search_exclude}{A character vector of flowsheet item names
(\code{disp_name} in raw format)  that are known NOT to be relevant to this
flowsheet item.}

\item{type}{Whether flowsheet item is a number or a string. Either
\code{"numeric"} or \code{"character"}}

\item{silently_exclude_na_when}{An expression that returns a logical value,
specifying the circumstances when \code{NA} values should be automatically
excluded.}

\item{silently_exclude_when}{An expression that returns a logical value,
specifying the circumstances when values should be automatically
excluded}

\item{censoring_fn}{A function specifying the value of the \code{censoring} column}

\item{value_as_number_fn}{A function specifying how to generate the
\code{value_as_number} column}

\item{value_as_character_fn}{A function specifying how to generate the
\code{value_as_character} column}

\item{value_as_logical_fn}{A function specifying how to generate the
\code{value_as_logical} column}

\item{unit_rescale_fn}{A function specifying how to rescale the flowsheet
values. By default, does not rescale}

\item{unit_relabel_fn}{A function specifying how to relabel the units of the
the flowsheet values. By default, does not rescale}

\item{coalesce_fn}{A function specifying how to combine measurements from
exactly the same time}

\item{expect_before}{An expression that returns a logical value,
specifying a condition that should be \code{TRUE} in the raw data}

\item{expect_after}{An expression that returns a logical value,
specifying a condition that should be \code{TRUE} in the processed (output) data}

\item{range_mainly_low, range_mainly_high}{An indicative lower and upper range
for most values. Values outside this range are NOT excluded: this is purely
for setting default scales of plots etc}

\item{range_discard_below, range_discard_above}{This pair set the lower and
upper range for the flowsheet item. Values outside this range are
EXCLUDED.}
}
\description{
Add a new definition of a flowsheet item
}
\author{
R.J.B. Goudie
}
