% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/during-fsheet.R
\name{fsheet_all_during}
\alias{fsheet_all_during}
\title{Left join with fsheet data}
\usage{
fsheet_all_during(x, y, during)
}
\arguments{
\item{x}{An \code{adm} data frame}

\item{y}{A (tidy) \code{fsheet} data frame: with \code{person_id}, \code{symbol},
\code{value_as_number} (or \code{value_as_character}; or both, plus a \code{type} column
indicating which to use when summarising) and \code{measurement_datetime}.}

\item{during}{The time period to extract data for, one of: \code{"during_visit"},
\code{"during_icu"}}
}
\value{
A data frame, with a row for each fsheet measurement during the relevant
time period for each patient. Note that rows of the \code{adm} data will be
repeated multiple times (one for each fsheet measurement data point).
}
\description{
Left join with fsheet data
}
\author{
R.J.B. Goudie
}
