% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsheet.R
\name{fsheet_annotate}
\alias{fsheet_annotate}
\title{Add annotations to fsheet data}
\usage{
fsheet_annotate(x, fsheet_def, annotation_db = NULL)
}
\arguments{
\item{x}{Flowsheet data in renamed format (after applying \code{fsheet_rename})}

\item{fsheet_def}{A fsheet definition, or list of fsheet definitions}

\item{annotation_db}{Either \code{NULL} or a \code{tbl} reference to the annotations
database table that the annotations should be stored in. If \code{NULL} the
annotations are added to \code{x} as additional columns.}
}
\description{
Annotate fsheet data to note rows that should be excluded etc.
}
\details{
Annotations can either be added as additional columns (when
\code{annotation_db = NULL}), or stored in a separate database table (by
setting \code{annotation_db} to a database \code{tbl}).

Note that when adding the annotations as additiona columns,
the result will not be sorted, since it is left in the original
order of \code{x}
}
\author{
R.J.B. Goudie
}
