% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsheet.R
\name{fsheet_extract}
\alias{fsheet_extract}
\title{Extract fsheet data into tidy format}
\usage{
fsheet_extract(
  x,
  fsheet_def,
  errors = stop,
  rds_only = FALSE,
  rds_filepath_fn = NULL
)
}
\arguments{
\item{x}{Flowsheet data in renamed format (after applying \code{fsheet_rename})}

\item{fsheet_def}{A fsheet definition}

\item{errors}{A function indicating what to do when an error is found}
}
\value{
A data frame with the following columns:
\code{person_id}, \code{symbol}, \code{value_as_character}, \code{value_as_number},
\code{value_as_logical}, \code{censoring}, \code{comment}, \code{measurement_datetime}, \code{name},
\code{title}, \code{data_id}, \code{measurement_id}, \code{line_id}, \code{template}, \code{form}, \code{type},
\code{unit}
}
\description{
Extract fsheet data into tidy format
}
\author{
R.J.B. Goudie
}
