% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsheet_io.R
\name{fsheet_io_unrename}
\alias{fsheet_io_unrename}
\title{Untidy raw fsheet_io colnames}
\usage{
fsheet_io_unrename(x)
}
\arguments{
\item{x}{A data frame of raw fsheet_io data with tidy names}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
Untidy raw fsheet_io colnames
}
\author{
R.J.B. Goudie
}
