% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsheet.R
\name{fsheet_nonnumeric}
\alias{fsheet_nonnumeric}
\title{Helper function to extract nonnumeric fsheet values}
\usage{
fsheet_nonnumeric(x)
}
\arguments{
\item{x}{A flowsheet data frame, with \code{value_as_number} column}
}
\value{
The rows of the supplied data frame that are \code{NA} after conversion to
to numeric form
}
\description{
Helper function to extract nonnumeric fsheet values
}
\author{
R.J.B. Goudie
}
