% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsheet.R
\name{fsheet_pivot_wider_datetime}
\alias{fsheet_pivot_wider_datetime}
\title{Pivot data wider, with multiple value type}
\usage{
fsheet_pivot_wider_datetime(
  x,
  id_cols = c("person_id", "measurement_datetime"),
  ...
)
}
\arguments{
\item{x}{A data frame}

\item{id_cols}{A set of columns that uniquely identifies each observation.}

\item{...}{Passed to \code{pivot_value_wider}}
}
\value{
A data frame
}
\description{
Pivot into \code{symbol} in columns, with each row a unique measurement time
}
\author{
R.J.B. Goudie
}
