% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsheet.R
\name{fsheet_rename}
\alias{fsheet_rename}
\title{Tidy raw fsheet colnames}
\usage{
fsheet_rename(x)
}
\arguments{
\item{x}{A data frame of raw fsheet data}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
Tidy raw fsheet colnames
}
\author{
R.J.B. Goudie
}
