% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsheet.R
\name{fsheet_spo2_on_room_air}
\alias{fsheet_spo2_on_room_air}
\title{Extract peripheral oxygen saturations (SpO2) on room air}
\usage{
fsheet_spo2_on_room_air(x, shape = "long")
}
\arguments{
\item{x}{A fsheet_raw dataframe}

\item{shape}{Either "long" or "wide" format dataframe output}
}
\description{
Excludes SpO2 measurements that are not from the same time as room air
}
