% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{geom_numeric_point_logical_point}
\alias{geom_numeric_point_logical_point}
\title{geom_numeric_point_logical_point}
\usage{
geom_numeric_point_logical_point()
}
\description{
geom_numeric_point_logical_point
}
