% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{grouped_summarise_or_slice}
\alias{grouped_summarise_or_slice}
\title{Summarise or slice grouped data}
\usage{
grouped_summarise_or_slice(x, type = "summarise", formula, group_by)
}
\arguments{
\item{x}{A data frame}

\item{type}{Either \code{"summarise"} or \code{"slice"}}

\item{formula}{An expression describing the summarise or slice}

\item{group_by}{A character vector listing columns that should be used
to group the summarise or slice}
}
\value{
A summarised (or sliced) data frame, usually with one row per group
}
\description{
Summarise or slice grouped data
}
\author{
R.J.B. Goudie
}
