% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inform_if_all_times_are_midnight}
\alias{inform_if_all_times_are_midnight}
\title{Alert user if all datetime are midnight}
\usage{
inform_if_all_times_are_midnight(x)
}
\arguments{
\item{x}{A \code{POSIXct} datetime}
}
\description{
Does nothing if a \code{Date} is supplied
}
