% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{inner_join_filter}
\alias{inner_join_filter}
\title{Inner join according to a filter and summarise in one}
\usage{
inner_join_filter(x, y, join_by, filter_by, filter_condition)
}
\arguments{
\item{x}{A data frame}

\item{y}{A data frame}

\item{join_by}{A character vector of variables to join \code{x} and \code{y} by}

\item{filter_by}{A character vector indicating the grouping to use when
filtering}

\item{filter_condition}{An expression that indicates whether the
corresponding row should be included}
}
\value{
A data frame, with all rows of \code{x} (potentially duplicated) with at
least one matching \code{y}
row, where the match is determined by joining by \code{join_by} and then
filtering using the \code{filter_condition}, under the \code{filter_by} grouping.
}
\description{
Inner join two data frames according to some columns (\code{join_by}), but then
remove some of the rows according to the \code{filter_condition} (which is run
after grouping via \code{filter_by}).
}
\author{
R.J.B. Goudie
}
