% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{inner_join_filter_pivot}
\alias{inner_join_filter_pivot}
\title{Inner join according to a filter and summarise in one}
\usage{
inner_join_filter_pivot(
  x,
  y,
  join_by,
  filter_by,
  filter_condition,
  type = "none",
  formula,
  pivot_by,
  names_from,
  values_from,
  values_fn_name = NULL
)
}
\arguments{
\item{x}{A data frame}

\item{y}{A data frame}

\item{join_by}{A character vector of variables to join \code{x} and \code{y} by}

\item{filter_by}{A character vector indicating the grouping to use when
filtering}

\item{filter_condition}{An expression that indicates whether the
corresponding row should be included}

\item{names_from, values_from}{A pair of arguments describing which column
(or columns) to get the name of the output column (\code{names_from}), and which
column (or columns) to get the cell values from (\code{values_from}).}

\item{values_fn_name}{A character string for naming the output of \code{values_fn}}
}
\value{
A data frame
}
\description{
Inner join according to a filter and summarise in one
}
\author{
R.J.B. Goudie
}
