% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{left_join_filter}
\alias{left_join_filter}
\title{Left join according to a filter and summarise in one}
\usage{
left_join_filter(x, y, ...)
}
\arguments{
\item{x}{A data frame}

\item{y}{A data frame}
}
\value{
A data frame, with all rows of \code{x}, potentially duplicated.
Where there is no matching \code{y} row, \code{NA}s will be present.

A match is determined by joining by \code{join_by} and then
filtering using the \code{filter_condition}, under the \code{filter_by} grouping.
}
\description{
Left join according to a filter and summarise in one
}
\author{
R.J.B. Goudie
}
