% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_admin.R
\name{med_admin_extract}
\alias{med_admin_extract}
\title{Extract med_admin data into tidy format}
\usage{
med_admin_extract(x, med_admin_def, errors = stop)
}
\arguments{
\item{x}{Medication administration data in tidy-column-name format (after
applying \code{med_admin_rename})}

\item{med_admin_def}{A med_admin definition}

\item{errors}{A function indicating what to do when an error is found}
}
\value{
A data frame with the following columns:
\code{person_id}, \code{symbol}, \code{value_as_number}, \code{value_as_character},
\code{censoring}, \code{administered_datetime},

\code{unit}, \code{name}, \code{title}, \code{method}, \code{group}, \code{type}

The result will be sorted by administered_datetime
}
\description{
Extract med_admin data into tidy format
}
\author{
R.J.B. Goudie
}
