% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_admin.R
\name{med_admin_rename}
\alias{med_admin_rename}
\title{Tidy raw med_admin colnames}
\usage{
med_admin_rename(x)
}
\arguments{
\item{x}{A data frame of raw med_admin data}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
Tidy raw med_admin colnames
}
\author{
R.J.B. Goudie
}
