% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_admin.R
\name{default_route_include}
\alias{default_route_include}
\alias{default_route_exclude}
\alias{default_action_include}
\alias{default_action_exclude}
\title{Route and actions defaults}
\usage{
default_route_include(route_class)

default_route_exclude(route_class)

default_action_include(action_class)

default_action_exclude(action_class)
}
\description{
Route and actions defaults
}
