% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_admin.R
\name{resolve_simultaneous_administration_datetime}
\alias{resolve_simultaneous_administration_datetime}
\alias{coalesce_simultaneous_administration_datetime}
\title{Resolve simultaneous administration}
\usage{
resolve_simultaneous_administration_datetime(x)

coalesce_simultaneous_administration_datetime(x)
}
\description{
Resolve simultaneous administration
}
