% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_admin.R
\name{med_admin_unrename}
\alias{med_admin_unrename}
\title{Untidy raw med_admin colnames}
\usage{
med_admin_unrename(x)
}
\arguments{
\item{x}{A data frame of raw med_admin data with tidy names}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
Untidy raw med_admin colnames
}
\author{
R.J.B. Goudie
}
