% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_hist.R
\name{med_hist_rename}
\alias{med_hist_rename}
\title{Tidy raw diagnosis_pl colnames}
\usage{
med_hist_rename(x)
}
\arguments{
\item{x}{A data frame of raw diagnosis_pl data}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
Tidy raw diagnosis_pl colnames
}
\author{
R.J.B. Goudie
}
