% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_prescr.R
\name{med_prescr_add}
\alias{med_prescr_add}
\alias{med_prescr_extract}
\title{Add a new definition of a test item}
\usage{
med_prescr_add(
  med_prescr_def,
  symbol,
  title,
  names = NA,
  search_pattern,
  search_exclude = NA,
  silently_exclude_na_routes = FALSE,
  route,
  route_exclude = NA,
  dose_as_number_fn = case_when(TRUE ~ dose_as_number),
  dose_rescale_fn = case_when(TRUE ~ as.numeric(dose_as_number)),
  unit_relabel_fn = case_when(TRUE ~ unit),
  expect_after = TRUE
)

med_prescr_extract(x, med_prescr_def, errors = stop)
}
\arguments{
\item{med_prescr_def}{A med_prescr definition}

\item{symbol}{A short name for the flowsheet item. This should be a
character vector that is suitable to use as an R object name}

\item{title}{The full name for the test item, suitable for using in
figure titles etc}

\item{names}{The names of the Medications (\code{DrugName} in raw format)}

\item{search_pattern}{A charater vector of potential synonymns for this
test, in case new names (\code{TestName} in raw format) for the
same test are added in future.}

\item{search_exclude}{A character vector of test names
(\code{TestName} in raw format) that are known NOT to be relevant to this test.}

\item{unit_relabel_fn}{A function specifying how to relabel the units of the
the result values. By default, does not rescale.}

\item{expect_after}{An expression that returns a logical value,
specifying a condition that should be \code{TRUE} in the processed (output) data}

\item{x}{Medication administration data in tidy-column-name format (after
applying \code{med_prescr_rename})}

\item{errors}{A function indicating what to do when an error is found}

\item{search_exclude_group}{A character vector of test group names
(\code{TestGroupName} in raw format) that are known NOT to be relevant to
this test.}

\item{type}{Whether flowsheet item is a number or a string. Either
\code{"numeric"} or \code{"character"}}

\item{silently_exclude_na_when}{An expression that returns a logical value,
specifying the circumstances when \code{NA} values should be automatically
excluded.}

\item{silently_exclude_when}{An expression that returns a logical value,
specifying the circumstances when values should be automatically
excluded}

\item{censoring_fn}{A function specifying the value of the \code{censoring} column}

\item{value_as_number_fn}{A function specifying how to generate the
\code{value_as_number} column}

\item{value_as_character_fn}{A function specifying how to generate the
\code{value_as_character} column}

\item{unit_rescale_fn}{A function specifying how to rescale the result
values. By default, does not rescale.}

\item{expect_before}{An expression that returns a logical value,
specifying a condition that should be \code{TRUE} in the raw data}

\item{range_mainly_low, range_mainly_high}{An indicative lower and upper range
for most values. Values outside this range are NOT excluded: this is purely
for setting default scales of plots etc}

\item{range_discard_below, range_discard_above}{This pair set the lower and
upper range for the flowsheet item. Values outside this range are
EXCLUDED.}
}
\value{
A data frame with the following columns:
\code{person_id}, \code{symbol}, \code{value_as_number}, \code{value_as_character},
\code{censoring}, \code{administered_datetime},

\code{unit}, \code{name}, \code{title}, \code{method}, \code{group}, \code{type}

The result will be sorted by administered_datetime
}
\description{
Add a new definition of a test item

Extract med_prescr data into tidy format
}
\author{
R.J.B. Goudie
}
