% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_prescr.R
\name{med_prescr_rename}
\alias{med_prescr_rename}
\title{Tidy raw med_prescr colnames}
\usage{
med_prescr_rename(
  x,
  names = c(person_id = "STUDY_SUBJECT_DIGEST", name = "DrugName", start_datetime =
    "StartDate", end_datetime = "EndDate", dose = "Dose", unit = "DoseUnit", strength =
    "Strength", dose_frequency = "DoseFrequency", route = "RouteOfMedication",
    in_or_out_patient = "InOrOutPatient", form = "FormOfMedication", thera_class =
    "THERA_CLASS", pharm_class = "PHARM_CLASS", pharm_subclass = "PHARM_SUBCLASS",
    visit_id = "PAT_ENC_CSN_ID", status = "OrderStatusCat", provider_type =
    "ProviderType", order_class = "Order_Class")
)
}
\arguments{
\item{x}{A data frame of raw med_prescr data}

\item{names}{A vector of new_name = old_name pairs}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument
}
\author{
R.J.B. Goudie
}
