% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_prescr.R
\name{med_prescr_stub_def}
\alias{med_prescr_stub_def}
\title{Generate a stub definition of medication administration}
\usage{
med_prescr_stub_def(x, symbol, title, search = NULL, names = NULL)
}
\arguments{
\item{x}{Medication data in tidy-column-name format (after applying
\code{med_prescr_rename})}

\item{symbol}{A short name for the medication administration item. This
should be a character vector that is suitable to use as an R object name}

\item{title}{The full name for the medication administration item, suitable
for using in figure titles etc}

\item{search}{A charater vector of potential synonymns for this
test, in case new names/variants of the same medication added in future.
If \code{NULL}, then \code{names} is used to define a specific list of medication
names to include.}

\item{names}{The names of the Medications (\code{DrugName} in raw format). Only
used if \code{search = NULL}.}
}
\description{
Returns the code needed to define a med_prescr_def object.
}
\details{
Typically, you may wish to run with \code{search} set to the drug name, and
then rerun once you have filtered the list of medications down. This
ensures that the \code{route} and \code{action} listed only includes those found
in the data for your medications of interest.
}
