% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_prescr.R
\name{med_prescr_unrename}
\alias{med_prescr_unrename}
\title{Untidy raw med_prescr colnames}
\usage{
med_prescr_unrename(
  x,
  names = c(STUDY_SUBJECT_DIGEST = "person_id", DrugName = "name", StartDate =
    "start_datetime", EndDate = "end_datetime", Dose = "dose", Strength = "strength",
    DoseUnit = "unit", DoseFrequency = "dose_frequency", RouteOfMedication = "route",
    InOrOutPatient = "in_or_out_patient", FormOfMedication = "form", THERA_CLASS =
    "thera_class", PHARM_CLASS = "pharm_class", PHARM_SUBCLASS = "pharm_subclass",
    PAT_ENC_CSN_ID = "visit_id", OrderStatusCat = "status", ProviderType =
    "provider_type", Order_Class = "order_class")
)
}
\arguments{
\item{x}{A data frame of raw med_prescr data with tidy names}

\item{names}{A vector of new_name = old_name pairs}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument
}
\author{
R.J.B. Goudie
}
