% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/during-functions.R
\name{nearest_to_event_during}
\alias{nearest_to_event_during}
\alias{nearest_visit_start_during}
\title{Extract measurements during a timeperiod nearest to a specific event}
\usage{
nearest_to_event_during(..., event_datetime, names_suffix = "")

nearest_visit_start_during(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=summarise_during]{summarise_during}}
  \describe{
    \item{\code{x}}{An \code{adm} data frame}
    \item{\code{y}}{A (tidy) data frame, with a \code{datetime} column}
    \item{\code{datetime}}{The column of \code{y} to use as the main datetime for matching}
    \item{\code{during}}{The time period to extract data for, one of:
\itemize{
\item \code{"during_visit"}
\item \code{"during_visit_initial_24h"}
\item \verb{"during_visit_initial_72h}
\item \code{"during_icu"}
\item \code{"before_visit_initial_24h"}
\item \code{"14_days_before_visit_until_visit_end"}
\item \code{"year_before_visit_until_visit_end"}
\item \code{"before_visit_end"}
\item \code{"30_days_before_visit_start"}
\item \code{"year_before_initial_24h"}
\item \code{"during_value"}
}}
  }}

\item{event_datetime}{The datetime that the extraction should be relative
to.}

\item{names_suffix}{A text label to use as the end of the column name
created.}
}
\description{
Extract the nearest measurement to a specific other event.
}
\examples{
#' # raw data that we will combine
#' # raw data that we will combine
adm_data_example
fsheet_news2_example

# extract the first NEWS2 record during each hospital visit
first_news2 <- adm_data_example \%>\%
  first_during(fsheet_news2_example,
             datetime = measurement_datetime,
             during = "during_visit") \%>\%
  select(person_id,
         visit_id,
         visit_start_datetime,
         visit_end_datetime,
         contains("news2"))

# extract the NEAREST (either before or after) SpO2 measurement that to
# the first NEWS2 record during each hospital visit
first_news2 \%>\%
 nearest_to_event_during(fsheet_spo2_example,
                         datetime = measurement_datetime,
                         during = "during_visit",
                         event_datetime = news2_first_during_visit_datetime,
                         names_suffix = "first_news2") \%>\%
  select(person_id,
         visit_id,
         contains("spo2"),
         contains("news2"))

# extract the NEAREST (either before or after) SpO2 measurement that to
# the first NEWS2 record during each hospital visit
first_news2 \%>\%
 nearest_visit_start_during(
   fsheet_spo2_example,
   datetime = measurement_datetime,
   during = "14_days_before_visit_until_visit_end") \%>\%
  select(person_id,
         visit_id,
         contains("spo2"),
         contains("news2"))
}
