% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nonnumeric}
\alias{nonnumeric}
\title{Helper function to extract nonnumeric value_as_numner values}
\usage{
nonnumeric(x)
}
\arguments{
\item{x}{A test data frame, with \code{value_as_number} column, after fixing
up various odd values etc}
}
\value{
The rows of the supplied data frame that are \code{NA} after conversion to
to numeric form
}
\description{
This tests whether everything that remains (after exclusions and conversion)
is numeric.
}
\author{
R.J.B. Goudie
}
