% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/during-summarise.R
\name{pivot_value_wider}
\alias{pivot_value_wider}
\title{Pivot data wider, with multiple value type}
\usage{
pivot_value_wider(
  x,
  id_cols = c("person_id", "visit_id"),
  names_from = "symbol",
  values_from = c("value_as_number", "value_as_character", "value_as_logical"),
  names_suffix = NULL
)
}
\arguments{
\item{x}{A data frame}

\item{id_cols}{A set of columns that uniquely identifies each observation.}

\item{names_from, values_from}{A pair of arguments describing which column
(or columns) to get the name of the output column (\code{group_by}), and which
column (or columns) to get the cell values from (\code{values_from}).}

\item{names_suffix}{A character string for naming the summarised output}
}
\value{
A data frame
}
\description{
Pivot long into a wider format, so that \code{names_from} appears in columns.
The function handles long format with a mix of \code{value_as_numeric},
\code{value_as_character} and \code{value_as_logical} value columns.
}
\details{
New columns are added to the end, are sorted (so that values and datetimes
are adjacent). The suffixes \verb{_value_as_number} and \verb{_value_as_character} are
also stripped to provide nicer column names.
}
\author{
R.J.B. Goudie
}
