% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tests_plot_numeric}
\alias{tests_plot_numeric}
\alias{tests_histogram_time_of_day}
\alias{tests_histogram_by_date}
\alias{tests_histogram_by_weekday}
\alias{tests_plot_point}
\alias{tests_plot_tile}
\title{Tests plots}
\usage{
tests_plot_numeric(
  x,
  range_mainly_low,
  range_mainly_high,
  range_discard_below,
  range_discard_above
)

tests_histogram_time_of_day(x)

tests_histogram_time_of_day(x)

tests_histogram_by_date(x)

tests_histogram_by_weekday(x)

tests_plot_point(x)

tests_plot_tile(x)
}
\description{
Tests plots
}
