% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc_report}
\alias{qc_report}
\title{QC report}
\usage{
qc_report(
  adm_data,
  adt_data,
  tests_data,
  tests_def,
  fsheet_data,
  fsheet_def,
  output_dir = getwd(),
  ...
)
}
\description{
QC report
}
\examples{
adm_data_example

# fsheet data
fsheet_def <- list() \%>\%
  fsheet_add(
    symbol = "news2",
    title = "NEWS2",
    names = c("NEWS2 score"),
    search_pattern = c("news2", "news"),
    search_exclude = c(),
    silently_exclude_na_when = TRUE,
    silently_exclude_when =
      (value_original == " " |
         value_original == "E" # some kind of typo
      ),
    value_as_number_fn =
      case_when(value_original == "3mm" ~ 3,
                TRUE ~ value_as_number),
    expect_after =
      (value_as_number \%in\% 0:17))
fsheet_news2_example

tests_def <- list() \%>\%
  tests_add(
    symbol = "magnesium",
    title = "Magnesium",
    names_cuh = c("MAGNESIUM"),
    names_external = "EXT MAGNESIUM",
    search_pattern = c("magnesium", "mg"),
    search_exclude = c("ETOH MG/100ML",
                       "RETIRED-ETOH MG/100ML",
                       "RIST 0.2MG/ML",
                       "RIST 0.77MG/ML",
                       "RISTOCETIN 0.5MG/ML",
                       "RISTOCETIN 1.5MG/ML"),
    silently_exclude_na_when = FALSE,
    censoring_fn = case_when(value_original == "<0.29" ~ "left", # very low
                             TRUE ~ NA_character_),
    value_as_number_fn = case_when(value_original == "<0.29" ~ 0.29,
                                   TRUE ~ value_as_number),
    expect_before = (unit == "mmol/L"),
    range_mainly_low = 0.4,
    range_mainly_high = 1.2,
    range_discard_below = 0.32,
    range_discard_above = 1.8)
tests_magnesium_example

\dontrun{
qc_report(adm_data = adm_data_example,
          adt_data = adt_data_example,
          tests_data = tests_magnesium_example,
          tests_def = tests_def,
          fsheet_data = fsheet_news2_example,
          fsheet_def = fsheet_def)
}
}
