% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiology.R
\name{radiology_extract}
\alias{radiology_extract}
\title{Extract radiology data into tidy format}
\usage{
radiology_extract(x, radiology_def, errors = stop)
}
\arguments{
\item{x}{Radiology data in tidy-column-name format (after
applying \code{radiology_rename})}

\item{radiology_def}{A radiology definition}

\item{errors}{A function indicating what to do when an error is found}
}
\value{
A data frame with the following columns:
\code{person_id}, \code{symbol}, \code{value_as_number}, \code{value_as_character},
\code{censoring}, \code{administered_datetime},

\code{unit}, \code{name}, \code{title}, \code{method}, \code{group}, \code{type}

The result will be sorted by administered_datetime
}
\description{
Extract radiology data into tidy format
}
\author{
R.J.B. Goudie
}
