% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiology.R
\name{radiology_rename}
\alias{radiology_rename}
\title{Tidy raw radiology colnames}
\usage{
radiology_rename(x)
}
\arguments{
\item{x}{A data frame of raw radiology data}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
Tidy raw radiology colnames
}
\author{
R.J.B. Goudie
}
