% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_tidybrookes_csv}
\alias{read_tidybrookes_csv}
\title{Load data from Epic in CSV format}
\usage{
read_tidybrookes_csv(
  file,
  col_types,
  n_max = Inf,
  na = c("", "NA"),
  quote = "\\"",
  tz = "Europe/London",
  nonexistent = NULL,
  ambiguous = NULL
)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data (either
a single string or a raw vector), as per \code{\link[readr:read_delim]{readr::read_delim()}}}

\item{col_types}{Either a string specifying the name of the table that the
data comes from, for which a standard format can be used. Available
value are \code{adm}, \code{adt}, \code{demogs}, \code{fsheet}, \code{fsheet_io}, \code{tests},
\code{med_admin}, \code{med_prescr}, \code{diagnosis_pl}, \code{med_hist}, \code{radiology}.

Or a \code{\link[readr:cols]{readr::cols()}} format list of column names and types, which
can be used where a nonstandard data format are supplied.}

\item{n_max}{Maximum number of lines to read.}

\item{na}{Character vector of strings to interpret as missing values.
Set this option to \code{character()} to indicate no missing values.}
}
\description{
Load data from Epic in CSV format
}
