% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidybrookes.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{select}
\alias{filter}
\alias{case_when}
\alias{if_else}
\alias{cols}
\alias{col_character}
\alias{col_datetime}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{case_when}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{select}}}

  \item{readr}{\code{\link[readr:parse_atomic]{col_character}}, \code{\link[readr:parse_datetime]{col_datetime}}, \code{\link[readr]{cols}}}
}}

