% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{relocate_and_clean_new_cols}
\alias{relocate_and_clean_new_cols}
\title{Tidy pivot_wider dataframe column names and arrangement}
\usage{
relocate_and_clean_new_cols(x, magic_prefix)
}
\arguments{
\item{x}{A data frame}

\item{magic_prefix}{A character string, providing the prefix that indicates
which columns are "new"}
}
\value{
The data frame \code{x}, which renamed and reorder columns
}
\description{
Moves new columns to the end, and sorts them (so that values and datetimes
are adjacent). The suffixes \verb{_value_as_number} and \verb{_value_as_character} are
also stripped to provide nicer column names
}
