% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relocate_ignoring_missing}
\alias{relocate_ignoring_missing}
\title{Change column order, ignoring missing columns}
\usage{
relocate_ignoring_missing(...)
}
\arguments{
\item{...}{A data frame, and rename pairs}
}
\description{
A version of \code{relocate} that gracefully handles columns that do not
exist in the supplied data frame. In this case, a message will inform users
of the missing columns, but the function will perform the relocate on
columns that exist without error.
}
