% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\alias{default_rename}
\alias{adm_rename}
\alias{adt_rename}
\alias{demogs_rename}
\alias{diagnosis_pl_rename}
\alias{fsheet_io_rename}
\alias{fsheet_rename}
\alias{med_admin_rename}
\alias{med_hist_rename}
\alias{med_prescr_rename}
\alias{radiology_rename}
\alias{procedures_rename}
\alias{tests_rename}
\title{Rename to tidy column names}
\usage{
default_rename(x)

adm_rename(x, names = default_rename("adm"))

adt_rename(x, names = c())

demogs_rename(x, names = c())

diagnosis_pl_rename(x, names = c())

fsheet_io_rename(x, names = c())

fsheet_rename(x, names = c())

med_admin_rename(x, names = c())

med_hist_rename(x, names = c())

med_prescr_rename(x, names = c())

radiology_rename(x, names = c())

procedures_rename(x, names = c())

tests_rename(x, names = c())
}
\arguments{
\item{x}{A data frame of raw adm data}

\item{names}{A vector of new_name = old_name pairs that will override
the default renaming (the default renaming is merged with these names,
with these names taking priority)}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
Rename columns from the default names in the standard data format from
Clinical Informatics into standardised, shorter names (e.g. all lowercase,
underscore-separated, with \verb{_datetime} suffix etc).
}
\details{
The default renaming map can be viewed using \code{default_rename()}.

If variations from this format occur, custom renaming can be performed
using the \code{names} argument.
}
\examples{
# The raw columns names are messy: long names, mixed capitalisation, etc!
fsheet_imported_example <-
  read_tidybrookes_csv(
    file = tidybrookes_example("fsheet.csv"),
    col_types = "fsheet"
  )
colnames(fsheet_imported_example)

# Default renaming can be viewed with
default_rename("fsheet")

colnames(fsheet_rename(fsheet_imported_example))

# Provided names override the defaults
fsheet_imported_example \%>\%
  fsheet_rename(names = c(custom_person_id = "STUDY_SUBJECT_DIGEST"))
}
\seealso{
\link{unrename} to reverse this process
}
\author{
R.J.B. Goudie
}
