% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_ignoring_missing}
\alias{rename_ignoring_missing}
\title{Rename columns, ignoring missing columns}
\usage{
rename_ignoring_missing(...)
}
\arguments{
\item{...}{A data frame, and rename pairs}
}
\description{
A version of \code{rename} that allows providing renaming pairs that do not
exist in the supplied data frame. In this case, a message will inform users
of the missing columns, but the function will performthe renaming on
columns that exist without error.
}
