% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{return_or_write_to_annotation_db}
\alias{return_or_write_to_annotation_db}
\title{Return of write to annotation database}
\usage{
return_or_write_to_annotation_db(x, x_annotated, annotation_db, id_cols)
}
\arguments{
\item{x}{The input \code{tbl} to the \verb{_annotate} function}

\item{x_annotated}{The out of the \verb{_annotate} function}

\item{annotation_db}{A \code{tbl} database source}

\item{id_cols}{A character vector of the columns to retain in the
annotations database (to enable linking)}
}
\description{
Return of write to annotation database
}
\keyword{internal}
