% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{summarise_pivot_wider}
\alias{summarise_pivot_wider}
\title{Inner join according to a filter and summarise in one}
\usage{
summarise_pivot_wider(
  x,
  type = "summary",
  formula,
  pivot_by,
  group_by,
  values_from,
  names_suffix = NULL
)
}
\arguments{
\item{x}{A data frame}

\item{group_by, values_from}{A pair of arguments describing which column
(or columns) to get the name of the output column (\code{group_by}), and which
column (or columns) to get the cell values from (\code{values_from}).}

\item{names_suffix}{A character string for naming the summary output}

\item{y}{A data frame}

\item{join_by}{A character vector of variables to join \code{x} and \code{y} by}

\item{filter_by}{A character vector indicating the grouping to use when
filtering}

\item{filter_condition}{An expression that indicates whether the
corresponding row should be included}
}
\value{
A data frame
Where \code{type} is \code{"none"} all matched rows in \code{y}.
Where \code{type} is \verb{"slice}, rows satifying the \code{formula} under \code{slice} are
returned
Where \code{type} is \code{"summarise"}, a new \code{summary} column is returned
}
\description{
Inner join according to a filter and summarise in one
}
\author{
R.J.B. Goudie
}
