% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{tbl_raw}
\alias{tbl_raw}
\title{Renamed view of raw data to tidy column names}
\usage{
tbl_raw(src, table, names = c())
}
\arguments{
\item{src}{A \code{DBIConnection} object produced by \link{duckdb_tidybrookes_connect}
or \code{DBI::dbConnect()}}

\item{table}{A string specifying the name of the table. Typical values
would be \code{adm}, \code{adt}, \code{demogs}, \code{fsheet}, \code{fsheet_io}, \code{tests}, \code{med_admin},
\code{med_prescr}, \code{diagnosis_pl}, \code{med_hist}, \code{radiology}}

\item{names}{A vector of new_name = old_name pairs}
}
\description{
Create view of a table with standardised, shorter column names (e.g. all
lowercase, underscore-separated, with \verb{_datetime} suffix etc) rather than
the default names in the standard data format from Clinical Informatics.
}
\details{
The default renaming map can be viewed using \code{default_rename()}.

If variations from this format occur, custom renaming can be performed
using the \code{names} argument.
}
\author{
R.J.B. Goudie
}
