% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{tbl_raw_unrenamed}
\alias{tbl_raw_unrenamed}
\title{Pointer to raw unrenamed table in duckdb}
\usage{
tbl_raw_unrenamed(src, table)
}
\arguments{
\item{src}{A \code{DBIConnection} object produced by \link{duckdb_tidybrookes_connect}
or \code{DBI::dbConnect()}}

\item{table}{A string specifying the name of the table. Typical values
would be \code{adm}, \code{adt}, \code{demogs}, \code{fsheet}, \code{fsheet_io}, \code{tests}, \code{med_admin},
\code{med_prescr}, \code{diagnosis_pl}, \code{med_hist}, \code{radiology}}
}
\description{
Pointer to raw unrenamed table in duckdb
}
