% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{tests_annotate}
\alias{tests_annotate}
\title{Add annotations to tests data}
\usage{
tests_annotate(x, tests_def, annotation_db = NULL)
}
\arguments{
\item{x}{Tests data in renamed format (after applying \code{tests_rename})}

\item{tests_def}{A tests definition, or list of tests definitions}

\item{annotation_db}{Either \code{NULL} or a \code{tbl} reference to the annotations
database table that the annotations should be stored in. If \code{NULL} the
annotations are added to \code{x} as additional columns.}
}
\description{
Annotate tests data to note rows that should be excluded etc.
}
\details{
Annotations can either be added as additional columns (when
\code{annotation_db = NULL}), or stored in a separate database table (by
setting \code{annotation_db} to a database \code{tbl}).

Note that when adding the annotations as additiona columns,
the result will not be sorted, since it is left in the original
order of \code{x}
}
\author{
R.J.B. Goudie
}
