% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{tests_info}
\alias{tests_info}
\title{Convert tests defintion list to a dataframe}
\usage{
tests_info(tests_def)
}
\arguments{
\item{tests_def}{A test definition list}
}
\description{
Convert tests defintion list to a dataframe
}
\examples{
tests_info(tests_def_example)
}
