% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{tests_nonnumeric}
\alias{tests_nonnumeric}
\title{Helper function to extract nonnumeric test values}
\usage{
tests_nonnumeric(x)
}
\arguments{
\item{x}{A test data frame, with \code{value} column}
}
\value{
The rows of the supplied data frame that are \code{NA} after conversion to
to numeric form
}
\description{
Helper function to extract nonnumeric test values
}
\author{
R.J.B. Goudie
}
