% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{tests_rename}
\alias{tests_rename}
\title{Tidy raw tests colnames}
\usage{
tests_rename(x)
}
\arguments{
\item{x}{A data frame of raw test data}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
Tidy raw tests colnames
}
\author{
R.J.B. Goudie
}
