% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{tests_unrename}
\alias{tests_unrename}
\title{Untidy raw tests colnames}
\usage{
tests_unrename(x)
}
\arguments{
\item{x}{A data frame of raw tests data with tidy names}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
Untidy raw tests colnames
}
\author{
R.J.B. Goudie
}
