% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidybrookes_example}
\alias{tidybrookes_example}
\title{Get path to tidybrookes example}
\usage{
tidybrookes_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
tidybrookes comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access
}
\examples{
tidybrookes_example()
tidybrookes_example("fsheet.csv")
}
