% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{wrap_def_if_single}
\alias{wrap_def_if_single}
\title{Wrap definition in a list}
\usage{
wrap_def_if_single(def)
}
\arguments{
\item{def}{A definition list}
}
\description{
Makes handling single-item definition lists easier by wrapping them in an
enclosing list, so that they can be handled using lapply etc.
}
