% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm.R
\name{adm_annotate}
\alias{adm_annotate}
\title{Annotate ADM data}
\usage{
adm_annotate(x)
}
\arguments{
\item{x}{A tidied adm data frame, as tidied by \code{\link[=adm_rename]{adm_rename()}}}
}
\value{
The supplied data frame \code{x} with additional annotations
}
\description{
Checks that all the gender data are in known values.
Adds \code{visit_length_days}, \code{person_id_short}.
}
\author{
R.J.B. Goudie
}
