% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm.R
\name{adm_rename}
\alias{adm_rename}
\title{Tidy raw tests colnames}
\usage{
adm_rename(x)
}
\arguments{
\item{x}{A data frame of raw adm data}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
Tidy raw tests colnames
}
\author{
R.J.B. Goudie
}
