% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm.R
\name{adm_rename}
\alias{adm_rename}
\title{Tidy raw tests colnames}
\usage{
adm_rename(
  x,
  names = c(person_id = "STUDY_SUBJECT_DIGEST", visit_id = "PAT_ENC_CSN",
    visit_start_datetime = "IN_DTTM", visit_end_datetime = "HOSP_DISCH_TIME", gender =
    "GENDER_DESC", ethnicity = "ETHNIC_GROUP_GROUPED", death_date = "DATE_OF_DEATH",
    age_at_visit_start = "AGE_AT_ADM", adm_service = "ADM_SERVICE", ward =
    "ADT_DEPARTMENT_NAME", discharge_destination = "DISCH_DEST", discharged_deceased =
    "DISCH_DECEASED", readmitted_within_30_days = "READMIT_WITHIN_30")
)
}
\arguments{
\item{x}{A data frame of raw adm data}

\item{names}{A vector of new_name = old_name pairs}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument
}
\author{
R.J.B. Goudie
}
