% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm.R
\name{adm_unrename}
\alias{adm_unrename}
\title{Untidy raw adm colnames}
\usage{
adm_unrename(
  x,
  names = c(STUDY_SUBJECT_DIGEST = "person_id", PAT_ENC_CSN = "visit_id", IN_DTTM =
    "visit_start_datetime", HOSP_DISCH_TIME = "visit_end_datetime", GENDER_DESC =
    "gender", ETHNIC_GROUP_GROUPED = "ethnicity", DATE_OF_DEATH = "death_date",
    AGE_AT_ADM = "age_at_visit_start", ADM_SERVICE = "adm_service", ADT_DEPARTMENT_NAME =
    "ward", DISCH_DEST = "discharge_destination", DISCH_DECEASED = "discharged_deceased",
    READMIT_WITHIN_30 = "readmitted_within_30_days")
)
}
\arguments{
\item{x}{A data frame of raw adm data with tidy names}

\item{names}{A vector of new_name = old_name pairs}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument
}
\author{
R.J.B. Goudie
}
