% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt.R
\name{adt_annotate}
\alias{adt_annotate}
\title{Basic cleanup, checks, and annotation of ADT data}
\usage{
adt_annotate(x, fixed_labels, annotate_fn)
}
\arguments{
\item{x}{Tidy ADT data, as tidied by \code{\link[=adt_rename]{adt_rename()}}}

\item{fixed_labels}{A data frame with a colum \code{department}, which will be
joined to the ADT data}

\item{annotate_fn}{A function that provides additional annotation}
}
\value{
The supplied tidy ADT data \code{x} with additional annotations
}
\description{
Performs basic cleanup, checks and annotation of ADT data with indexes for the
hospital visit (\code{visit_index}) and the visit to each departments
(\code{department_visit_index}), and basic labels for each \code{department}.
}
\details{
The clean and checks performed are:
\enumerate{
\item Removes duplicate rows
\item Checks discharge dates are identical within a visit
}

The annotations added are:
\enumerate{
\item Labels hospital visits via \code{visit_index}
\item Adds the \code{end_datetime} for each row of ADT (which is not always included
in the raw data)
\item Labels visits to specific departments. A visit to a single department
may involve multiple rows of ADT data, since a patient may move beds
within a department one or more times. The visits are labelled with
\code{department_visit_index}, and the corresponding times recorded in
\code{department_start_datetime} and \code{department_end_datetime}, and
\code{department_length_days}.
\item (Left) joins in \code{fixed_labels}
\item Runs the user-supplied annotation function \code{annotate_fn}
}
}
\examples{
fixed_labels <-
  tibble::tribble(
    ~department, ~is_ed,
    "ADD EMERGENCY DEPT", TRUE,
    "ADD C2 WARD", FALSE,
    "ADD D2 WARD", FALSE,
    "DIALYSIS CENTRE", FALSE,
    "ADD A1 WARD", FALSE,
    "ADD A2 WARD", FALSE,
    "ADD A3 WARD", FALSE,
    "ADD A4 WARD", FALSE,
    "ADD X WARD", FALSE,
    "ADD Y WARD", FALSE,
    "ADD Z WARD", FALSE)

adt_data_example <-
  read_tidybrookes_csv(
    file = tidybrookes_example("adt.csv"),
    col_types = "adt"
  ) \%>\%
  adt_rename

adt_data_example \%>\%
  adt_annotate(annotate_fn = identity,
               fixed_labels = fixed_labels)

}
\seealso{
\code{\link[=adm_annotate]{adm_annotate()}} annotates ADM data
}
\author{
R.J.B. Goudie
}
