% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt.R
\name{adt_annotate}
\alias{adt_annotate}
\title{Annotate ADT data}
\usage{
adt_annotate(x, fixed_labels, annotate_fn)
}
\arguments{
\item{x}{Tidy ADT data, as tidied by \code{\link[=adt_rename]{adt_rename()}}}

\item{fixed_labels}{A data frame with a colum \code{department}, which will be
joined to the ADT data}

\item{annotate_fn}{A function that provides additional annotation}
}
\value{
The supplied tidy ADT data \code{x} with additional annotations
}
\description{
Removes duplicate rows
}
\details{
Preforms the following checks
\itemize{
\item Checks discharge dates are identical within a visit
}

Annotates the ADT data with
\itemize{
\item \code{visit_index}
\item \code{end_datetime}
}
}
\author{
R.J.B. Goudie
}
