% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt.R
\name{adt_check_discharge_dates_consistent}
\alias{adt_check_discharge_dates_consistent}
\title{Check disharge date consistent with final date}
\usage{
adt_check_discharge_dates_consistent(x)
}
\description{
Check disharge date consistent with final date
}
