% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt.R
\name{adt_check_discharge_dates_identical}
\alias{adt_check_discharge_dates_identical}
\title{Check discharge dates identical within a visit}
\usage{
adt_check_discharge_dates_identical(x)
}
\arguments{
\item{x}{Tidy ADT data, as tidied by \code{\link[=adt_rename]{adt_rename()}}}
}
\value{
The supplied tidy ADT data \code{x}, with an additional logical column
\code{check_all_discharge_dates_identical}, with \code{FALSE} indicating
inconsistencies
}
\description{
Check that all discharge_datetime are the same within a `visit_id``
}
\author{
R.J.B. Goudie
}
