% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/during-all.R
\name{all_during}
\alias{all_during}
\title{Join adm data with other (standardised) data}
\usage{
all_during(
  x,
  y,
  datetime,
  during,
  names_from = "symbol",
  join = "inner",
  arrange = TRUE
)
}
\arguments{
\item{x}{An \code{adm} data frame}

\item{y}{A (tidy) data frame, with a \code{datetime} column}

\item{datetime}{The column of \code{y} to use as the main datetime for matching}

\item{during}{The time period to extract data for, one of:
\itemize{
\item \code{"during_visit"}
\item \code{"during_visit_initial_24h"}
\item \verb{"during_visit_initial_72h}
\item \code{"during_icu"}
\item \code{"before_visit_initial_24h"}
\item \code{"14_days_before_visit_until_visit_end"}
\item \code{"year_before_visit_until_visit_end"}
\item \code{"before_visit_end"}
\item \code{"30_days_before_visit_start"}
\item \code{"year_before_initial_24h"}
\item \code{"during_value"}
}}

\item{names_from}{A character vector containing the names of columns that
should be used to filter by (in addition to those implied by \code{during}).}

\item{join}{Either \code{"left"}, in which a left join is performed (so all \code{x}
rows will be returned, even if they have no correspoding \code{y}) or
\code{"inner"}, in which an inner join is performed (so only \code{x} rows with at
least one \code{y} row will be returned)}
}
\value{
A data frame, with a row for each \code{y} measurement during the relevant
time period for each patient. Note that rows of the \code{adm} data will be
repeated multiple times (one for each \code{y} measurement data point).
}
\description{
Joins an \code{adm} data frame with another data frame (with a
\code{person_id} for linking and with the relevant datetime called \code{datetime}),
and filter to only those rows of \code{y} with \code{datetime} within the specified
time period.
}
\examples{
adm_data_example
fsheet_news2_example

all_during(adm_data_example,
           fsheet_news2_example,
           datetime = measurement_datetime,
           during = "during_visit") \%>\%
  select(person_id, visit_id, symbol, measurement_datetime, value_as_number)
}
\seealso{
If only a summary of the values during a timeperiod is needed,
then \code{\link[=summarise_during]{summarise_during()}} will be more convenient, or the more
specialised forms described at \link{during_functions}, \link{adjacent_event_during},
\link{extremes_during}.
}
\author{
R.J.B. Goudie
}
