% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{all_during}
\alias{all_during}
\title{Left join adm data with other (standardised) data}
\usage{
all_during(x, y, during, names_from)
}
\arguments{
\item{x}{An \code{adm} data frame}

\item{y}{A (tidy) data frame, with a \code{datetime} column}

\item{during}{The time period to extract data for, one of: \code{"during_visit"},
\code{"during_icu"}}
}
\value{
A data frame, with a row for each \code{y} measurement during the relevant
time period for each patient. Note that rows of the \code{adm} data will be
repeated multiple times (one for each \code{y} measurement data point).
}
\description{
Left joins an \code{adm} data frame with another data frame (with a
\code{person_id} for linking and with the relevant datetime called \code{datetime}),
and filter to only those rows of \code{y} with \code{datetime} within the specified
time period.
}
\author{
R.J.B. Goudie
}
