% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{condition_add}
\alias{condition_add}
\alias{condition_extract}
\title{Add a new definition of a Diagnosis PL item}
\usage{
condition_add(
  condition_def,
  symbol,
  title,
  icd10 = NA,
  condition_start_datetime_fn = case_when(!is.na(diagnosis_datetime) ~
    diagnosis_datetime, !is.na(entered_datetime) ~ entered_datetime, TRUE ~ ymd_hms(NA)),
  condition_end_datetime_fn = case_when(!is.na(resolved_datetime) ~ resolved_datetime,
    TRUE ~ ymd_hms(NA)),
  silently_exclude_deleted_when = TRUE,
  silently_exclude_resolved_when = FALSE
)

condition_extract(x, condition_def, errors = stop)
}
\arguments{
\item{condition_def}{A condition definition}

\item{symbol}{A short name for the condition item. This should be a
character vector that is suitable to use as an R object name}

\item{title}{The full name for the condition item, suitable for using in
figure titles etc}

\item{x}{condition data in tidy-column-name format (after applying
\code{condition_rename})}

\item{errors}{A function indicating what to do when an error is found}

\item{icd_10}{A vector of ICD-10 codes to use.}

\item{silently_exclude_deleted}{When TRUE, diagnoses with
\code{status = "Deleted"} will be removed from the data frame. In theory, only
diagnoses made in error should be deleted, but in practice it is thought
that clinical staff often delete diagnoses that are resolved rather than
mark them as "Resolved".}
}
\value{
A data frame with the following columns:
\code{person_id}, \code{symbol}, \code{value_as_number}, \code{value_as_character},
\code{censoring}, \code{ordered_datetime}, \code{collected_datetime}, \code{received_datetime},
\code{result_datetime}, \code{ordering_department_name}, \code{range_low}, \code{range_high},
\code{unit}, \code{name}, \code{title}, \code{method}, \code{group}, \code{type}

The result will be sorted by diagnosis_datetime
}
\description{
Add a new definition of a Diagnosis PL item

Extract condition data into tidy format
}
\author{
R.J.B. Goudie
}
