% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/during-functions.R
\name{first_during}
\alias{first_during}
\alias{last_during}
\alias{nearest_visit_start_during}
\alias{last_during_before_event}
\title{During functions}
\usage{
first_during(...)

last_during(...)

nearest_visit_start_during(...)

last_during_before_event(..., event_datetime, names_suffix = "")
}
\description{
During functions
}
