% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/during-functions.R
\name{first_during}
\alias{first_during}
\alias{last_during}
\title{Extract first or last measurement during a timeperiod}
\usage{
first_during(...)

last_during(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=summarise_during]{summarise_during}}
  \describe{
    \item{\code{x}}{An \code{adm} data frame}
    \item{\code{y}}{A (tidy) data frame, with a \code{datetime} column}
    \item{\code{datetime}}{The column of \code{y} to use as the main datetime for matching}
    \item{\code{during}}{The time period to extract data for, one of:
\itemize{
\item \code{"during_visit"}
\item \code{"during_visit_initial_24h"}
\item \verb{"during_visit_initial_72h}
\item \code{"during_icu"}
\item \code{"before_visit_initial_24h"}
\item \code{"14_days_before_visit_until_visit_end"}
\item \code{"year_before_visit_until_visit_end"}
\item \code{"before_visit_end"}
\item \code{"30_days_before_visit_start"}
\item \code{"year_before_initial_24h"}
\item \code{"during_value"}
}}
  }}
}
\description{
Extract the first measurement or the last measurement during a timeperiod.
}
\examples{
# raw data that we will combine
adm_data_example
fsheet_news2_example

# extract the first of each fsheet symbol during the hospital visit
adm_data_example \%>\%
  first_during(fsheet_news2_example,
               datetime = measurement_datetime,
               during = "during_visit") \%>\%
  select(person_id,
         visit_id,
         visit_start_datetime,
         visit_end_datetime,
         contains("news2"))

# extract the last of each fsheet symbol during the hospital visit
adm_data_example \%>\%
 last_during(fsheet_news2_example,
             datetime = measurement_datetime,
             during = "during_visit") \%>\%
  select(person_id,
         visit_id,
         visit_start_datetime,
         visit_end_datetime,
         contains("news2"))

# can pipe together two summaries
adm_data_example \%>\%
  first_during(fsheet_news2_example,
               datetime = measurement_datetime,
               during = "during_visit") \%>\%
  last_during(fsheet_news2_example,
              datetime = measurement_datetime,
              during = "during_visit") \%>\%
  select(person_id,
         visit_id,
         visit_start_datetime,
         visit_end_datetime,
         contains("news2"))

# example data with 3 fsheet data fields
fsheet_example

# summaries for all symbols can be calculated together
adm_data_example \%>\%
  first_during(fsheet_example,
               datetime = measurement_datetime,
               during = "during_visit") \%>\%
  select(person_id,
         visit_id,
         visit_start_datetime,
         visit_end_datetime,
         contains("during"))
}
