% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsheet.R
\name{fsheet_annotate}
\alias{fsheet_annotate}
\title{Add annotations to fsheet data}
\usage{
fsheet_annotate(x, fsheet_def, annotation_db = NULL)
}
\arguments{
\item{x}{Flowsheet data in renamed format (after applying \code{fsheet_rename})}

\item{fsheet_def}{A fsheet definition, or list of fsheet definitions}

\item{annotation_db}{Either \code{NULL} or a \code{tbl} reference to the annotations
database table that the annotations should be stored in. If \code{NULL} the
annotations are added to \code{x} as additional columns.}
}
\description{
Add annotations to fsheet data
}
\author{
R.J.B. Goudie
}
