% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsheet.R
\name{fsheet_add}
\alias{fsheet_add}
\alias{fsheet_extract}
\title{Add and extract flowsheet items}
\usage{
fsheet_add(
  fsheet_def,
  symbol,
  title,
  names = NA,
  search_pattern,
  search_exclude = NA,
  type = "numeric",
  silently_exclude_na_when = FALSE,
  silently_exclude_when = FALSE,
  censoring_fn = case_when(TRUE ~ NA_character_),
  value_as_number_fn = if_else(type == "numeric", case_when(TRUE ~ value_as_number),
    case_when(TRUE ~ NA_real_)),
  value_as_character_fn = if_else(type == "character", case_when(TRUE ~
    value_as_character), case_when(TRUE ~ NA_character_)),
  value_as_logical_fn = if_else(type == "logical", case_when(TRUE ~ value_as_logical),
    case_when(TRUE ~ as.logical(NA))),
  unit_rescale_fn = case_when(TRUE ~ value_as_number),
  unit_relabel_fn = case_when(TRUE ~ NA_character_),
  coalesce_fn = identity,
  expect_before = TRUE,
  expect_after = TRUE,
  range_mainly_low = NA,
  range_mainly_high = NA,
  range_discard_below = NA,
  range_discard_above = NA
)

fsheet_extract(
  x,
  fsheet_def,
  errors = stop,
  rds_only = FALSE,
  rds_filepath_fn = NULL
)
}
\arguments{
\item{fsheet_def}{A list of fsheet definitions}

\item{symbol}{A short name for the flowsheet item. This should be a character
vector that is suitable to use as an R object name}

\item{title}{The full name for the flowsheet item, suitable for using in
figure titles etc}

\item{names}{A character vector of display names (\code{disp_name} in raw format)
of the flowsheet items to include}

\item{search_pattern}{A character vector of potential synonymns for this
flowsheet item, in case new names (\code{disp_name} in raw format) for the
same item are added in future. These will be searched for in the raw data
and if any potential new names are identified, the user will be alerted to
the presence of potential new flowsheet names.}

\item{search_exclude}{A character vector of flowsheet item names
(\code{disp_name} in raw format) that are known NOT to be relevant to this
flowsheet item.}

\item{type}{Whether flowsheet item is a number or a string. Either
\code{"numeric"} or \code{"character"}}

\item{silently_exclude_na_when}{An expression that returns a logical value,
specifying the circumstances when \code{NA} values should be automatically
excluded.}

\item{silently_exclude_when}{An expression that returns a logical value,
specifying the circumstances when values should be automatically
excluded}

\item{censoring_fn}{A function specifying the value of the \code{censoring} column}

\item{value_as_number_fn}{A function specifying how to generate the
\code{value_as_number} column}

\item{value_as_character_fn}{A function specifying how to generate the
\code{value_as_character} column}

\item{value_as_logical_fn}{A function specifying how to generate the
\code{value_as_logical} column}

\item{unit_rescale_fn}{A function specifying how to rescale the flowsheet
values. By default, does not rescale}

\item{unit_relabel_fn}{A function specifying how to relabel the units of the
the flowsheet values. By default, does not rescale}

\item{coalesce_fn}{A function specifying how to combine measurements from
exactly the same time}

\item{expect_before}{An expression that returns a logical value,
specifying a condition that should be \code{TRUE} in the raw data}

\item{expect_after}{An expression that returns a logical value,
specifying a condition that should be \code{TRUE} in the processed (output) data}

\item{range_mainly_low, range_mainly_high}{An indicative lower and upper range
for most values. Values outside this range are NOT excluded: this is purely
for setting default scales of plots etc}

\item{range_discard_below, range_discard_above}{This pair set the lower and
upper range for the flowsheet item. Values outside this range are
EXCLUDED.}

\item{x}{Flowsheet data in renamed format (after applying \code{fsheet_rename})}

\item{errors}{A function indicating what to do when an error is found}
}
\value{
\code{fsheet_add} returns a list of fsheet definitions, with the new definition
added to the end of the supplied fsheet_def

\code{fsheet_extract} returns a data frame with the following columns:
\code{person_id}, \code{symbol}, \code{value_as_character}, \code{value_as_number},
\code{value_as_logical}, \code{censoring}, \code{comment}, \code{measurement_datetime}, \code{name},
\code{title}, \code{data_id}, \code{measurement_id}, \code{line_id}, \code{template}, \code{form}, \code{type},
\code{unit}
}
\description{
\code{fsheet_add} creates a new "fsheet definition", which is a specification of
how the raw flowsheet data should be extracted, filtered, cleaned and
standardised. A list of such specifications can be used on raw \code{fsheet} data
using \code{fsheet_extract}.
}
\details{
Flowsheet defintiions describe how the raw data should be extracted,
filtered, cleaned and standardised as follows:
\enumerate{
\item \strong{Labelling the flowsheet item}. Two names/labels are required each
fsheet item:
\itemize{
\item a \code{symbol}, which serves as the "R variable name" for the item. To make
the symbol quick and easy to type, and work with in R, it is
recommended that these are lowercase, ideally single words (e.g.
\code{"news2"} or "\code{spo2})
\item a \code{title}, which serves as the label the of the flowsheet item in
plots and figures. This should be capitalised as you would want the
item to be labelled in formal documents (e.g. \code{"NEWS2"} or \verb{"SpO2}).
}
\item \strong{Extracting relevant flowsheet items} The specification of which
rows should be extracted is via:
\itemize{
\item \code{names}, a vector listing the labels of the fsheet items in the raw
data. This corresponds to column labelled as \code{name} by
\code{\link{fsheet_rename}}, which is labelled as \code{disp_name} in the raw
extracts from Epic).
}

Additionally, to guard against the potential for new names for flowsheet
items being introduced through the evolution of the EHR data, search terms
for potential other names for this flowsheet items (and known exclusions)
can be provided using:
\itemize{
\item \code{search_pattern}, a vector of search terms of potential synonyms for
the flowsheet item. This list of search terms will be searched
whenever the flowsheet item is extracted, and if a possible new data
field is identified, then this will be alerted to the user.
\item \code{search_exclude}, a vector of names of flowsheet items that are known
\emph{not} to be relevant. This list of names will be removed prior to
searching the data for the search patterns in \code{search_pattern}.
This helps to avoid alerting the user about numerous flowsheet items
that are known to be irrelevant.
}
\item \strong{Excluding rows}. Rows containing \code{NA} in the \code{value} column may be
always irrelevant, so it may be helpful to remove these by setting
\code{silently_exclude_na_when} to \code{TRUE}.

Alternatively, an expression that describes when row should be excluded
can be provided using \code{silently_exclude_when}. For example, setting this
to \code{(value_original == "Other (Comment)")} will exclude all rows where the
value is \code{"Other (Comment)"}.

For numerical flowsheet items,\code{range_discard_below} and
\code{range_discard_above} can be specified so that values outside this range
are excluded (see below for details).
\item \strong{Converting values to numbers/characters/logicals}. Raw flowsheet data
are a complex mix of numerical data (sometimes containing censoring, and
other non-numerical data), categorical data (coded as characters) and
logical data (usually coded as characters). Standardising the raw data
into data of a single type makes downstream analysis easier, so several
aspects of the definition specify how this is done:
\itemize{
\item \code{type}: one of \code{numeric}, \code{character} or \code{logical}. This is used to
set sensible defaults for other aspects of the specification, and is
added as a column to the output, so downstream handling of the data
can use this column to determine handling of the data (e.g. default
plots might be a barplot for \code{character} items and a lineplot for
\code{numeric} items)
\item \code{value_as_number_fn}, \code{value_as_character_fn} and \code{value_as_logical_fn}
enable functions that handle conversion of the raw data.
\strong{TODO add more details}
\item \code{unit_rescale_fn} and \code{unit_relabel_fn} enable functions that standardise
the units to be provided
\item \code{coalesce_fn} enables a function that combines together data to be
provided. For instance, this can be used to specify how simultaneous data
should be combined: e.g. how should two blood pressure records at
the exact same time be handled? The function can describe how to combine
the information by e.g. taking the mean or max, or arbitrarily choosing
one of the records.
}
\item \strong{Validating that data conform to expectations}. Two parameters enable
checking that the data comply with known restrictions. For example, we
may know a finite range exists for a flowsheet item, and we want to check
that all data fall within this range.

Alternatively, we may need to map each value of the raw data to a smaller
range of values, and we need to be sure that our function for doing this
handles all the values that are present in the raw data (i.e. no new
values have been added).

Checks that should be performed on the raw data should be supplied as
\code{expect_before}, whereas checks that should be performed on the cleaned
data should be supplied as \code{expect_after}.
\item \strong{Specification of ranges}. Two types of ranges can be specified:
\itemize{
\item \code{range_discard_below} and \code{range_discard_above}: these are intended to
describe \emph{hard} limits: values outside this range will be \emph{excluded}.
\item \code{range_mainly_low} and \code{range_mainly_high}: these are intended to
describe the range of values we typically expect most of the data to
fall within, which can be useful for setting default ranges for e.g.
default plots. Values outside this range are \emph{not excluded}.
}
}
}
\examples{
# A definition for the NEWS2 score:
# - Includes the value of the flowsheet called "NEWS 2 score"
# - Excludes all NA values, and also " " and "E".
# - The value_as_number_fn will convert "3mm" to 3, since this is probably
#   a typo
# - Checks that all output values are in integers between 0 and 17.
fsheet_def <- list() \%>\%
  fsheet_add(
    symbol = "news2",
    title = "NEWS2",
    names = c("NEWS2 score"),
    search_pattern = c("news2", "news"),
    search_exclude = c(),
    silently_exclude_na_when = TRUE,
    silently_exclude_when =
      (value_original == " " |
         value_original == "E" # some kind of typo
      ),
    value_as_number_fn =
      case_when(value_original == "3mm" ~ 3,
                TRUE ~ value_as_number),
    expect_after =
      (value_as_number \%in\% 0:17))

# Note that person_id == "BB" has a NEWS2 with value="3mm", and that
# person_id == "CC" has a NEWS2 with NA value
fsheet_raw_example \%>\% print(n = Inf)

# Note that the NA row is excluded, and that "3mm" is converted to 3
fsheet_extract(fsheet_raw_example, fsheet_def)
}
\author{
R.J.B. Goudie
}
