% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsheet_io.R
\name{fsheet_io_rename}
\alias{fsheet_io_rename}
\title{Tidy raw fsheet_io colnames}
\usage{
fsheet_io_rename(x, names = default_rename("fsheet_io"))
}
\arguments{
\item{x}{A data frame of raw fsheet_io data}

\item{names}{A vector of new_name = old_name pairs}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument
}
\author{
R.J.B. Goudie
}
