% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsheet_io.R
\name{fsheet_io_unrename}
\alias{fsheet_io_unrename}
\title{Untidy raw fsheet_io colnames}
\usage{
fsheet_io_unrename(
  x,
  names = c(STUDY_SUBJECT_DIGEST = "person_id", disp_name = "description",
    FLO_MEAS_NAME = "name", measured_value = "value", meas_comment = "comment",
    MEASURE_TIME = "measurement_datetime", fsd_id = "data_id", `flo-meas_id` =
    "measurement_id", line = "line_id", template = "template", form = "form")
)
}
\arguments{
\item{x}{A data frame of raw fsheet_io data with tidy names}

\item{names}{A vector of new_name = old_name pairs}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument
}
\author{
R.J.B. Goudie
}
