% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsheet.R
\name{fsheet_unrename}
\alias{fsheet_unrename}
\title{Untidy raw fsheet colnames}
\usage{
fsheet_unrename(x)
}
\arguments{
\item{x}{A data frame of raw fsheet data with tidy names}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
Untidy raw fsheet colnames
}
\author{
R.J.B. Goudie
}
