% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{inner_join_filter}
\alias{inner_join_filter}
\title{Inner join according to a filter and summarise in one}
\usage{
inner_join_filter(x, y, join_by, filter_by, filter_condition, type, formula)
}
\arguments{
\item{x}{A data frame}

\item{y}{A data frame}

\item{join_by}{A character vector of variables to join \code{x} and \code{y} by}

\item{filter_by}{A character vector indicating the grouping to use when
filtering}

\item{filter_condition}{An expression that indicates whether the
corresponding row should be included}

\item{type}{Either \code{"slice"}, \verb{"summarise}, or \verb{"none} (for to return all
values)}

\item{formula}{An expression for slicing or summarising}
}
\value{
A data frame, with all rows of \code{x} (potentially duplicated) with at
least one matching \code{y}
row, where the match is determined by joining by \code{join_by} and then
filtering using the \code{filter_condition}, under the \code{filter_by} grouping.

Where \code{type} is \code{"none"} all matched rows in \code{y}.
Where \code{type} is \verb{"slice}, rows satifying the \code{formula} under \code{slice} are
returned
Where \code{type} is \verb{"summarise}, a new \code{summary} column is returned
}
\description{
Inner join according to a filter and summarise in one
}
\author{
R.J.B. Goudie
}
