% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_admin.R
\name{med_admin_add}
\alias{med_admin_add}
\alias{med_admin_extract}
\title{Add a new definition of a medication administration drug}
\usage{
med_admin_add(
  med_admin_def,
  note,
  symbol,
  title,
  names = NA,
  search_pattern,
  search_exclude = NA,
  silently_exclude_na_routes = FALSE,
  route_class = "intravenous",
  route_include = default_route_include(route_class),
  route_exclude = default_route_exclude(route_class),
  action_class = c("bolus", "infusion", "pca"),
  action_include = default_action_include(action_class),
  action_exclude = default_action_exclude(action_class),
  concentration_fn = case_when(TRUE ~ NA_real_),
  dose_range_discard_above_fn = case_when(TRUE ~ Inf),
  dose_range_mainly_below_fn = NA,
  expect_after = TRUE
)

med_admin_extract(x, med_admin_def, med_admin_units, errors = stop)
}
\arguments{
\item{med_admin_def}{A med_admin definition}

\item{note}{A character vector of notes that should be printed}

\item{symbol}{A short name for the flowsheet item. This should be a
character vector that is suitable to use as an R object name}

\item{title}{The full name for the test item, suitable for using in
figure titles etc}

\item{names}{The names of the Medications (\code{DrugName} in raw format)}

\item{search_pattern}{A charater vector of potential synonymns for this
drug, in case new names (\code{DrugName} in raw format) for the
same drug are added in future.}

\item{search_exclude}{A character vector of drug names
(\code{DrugName} in raw format) that are known NOT to be relevant to this drug}

\item{silently_exclude_na_routes}{An logical specifying whether medication
administration records with \code{NA} routes should be automatically
excluded.}

\item{route_class}{One of \code{enteral}, \code{intravenous}, \code{other_systemic},
\code{respiratory} or \code{topical}. These classes will use the
classification in \code{med_admin_routes.csv} to specify a list of routes
to include}

\item{route_include}{A vector of route names (\code{RouteOfMedicationAbbreviated}
in raw format) to include}

\item{route_exclude}{A vector of route names (\code{RouteOfMedicationAbbreviated}
in raw format) to exclude}

\item{action_class}{One of \code{infusion}, \code{bolus}, \code{pca}, \code{feed}. These classes
will use the classification in \code{med_admin_action.csv} to specify a list of
actions to include}

\item{action_include}{A vector of action names (\code{MARAction} in raw format) to
include}

\item{action_exclude}{A vector of action names (\code{MARAction} in raw format) to
exclude}

\item{concentration_fn}{A function that codes the concentration of each
row of the medication administration records (for that drug). It should
return a real. This will typically be done on the basis of
the \code{name}, \code{info_concentration}, \code{info_conversion} or \code{info_rate},
but can use any variables in the dataframe.}

\item{dose_range_mainly_below_fn}{NOT CURRRENTLY IMPLEMENTED.
An indicative lower and upper range
for most values. Values outside this range are NOT excluded: this is purely
for setting default scales of plots etc}

\item{expect_after}{An expression that returns a logical value,
specifying a condition that should be \code{TRUE} in the processed (output) data}

\item{x}{Medication administration data in tidy-column-name format (after
applying \code{med_admin_rename})}

\item{errors}{A function indicating what to do when an error is found}

\item{dose_range_discard_above_fne}{NOT CURRRENTLY IMPLEMENTED.
This seta the upper range for the dose item. Values outside this range are
EXCLUDED.}
}
\value{
A data frame with the following columns:
\code{person_id}, \code{symbol}, \code{value_as_number}, \code{value_as_character},
\code{censoring}, \code{administered_datetime},

\code{unit}, \code{name}, \code{title}, \code{method}, \code{group}, \code{type}

The result will be sorted by administered_datetime
}
\description{
Add a new definition of a medication administration drug

Extract med_admin data into tidy format
}
\author{
R.J.B. Goudie
}
