% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_admin.R
\name{med_admin_rename}
\alias{med_admin_rename}
\title{Tidy raw med_admin colnames}
\usage{
med_admin_rename(
  x,
  names = c(person_id = "STUDY_SUBJECT_DIGEST", administered_datetime =
    "TimeAdministered", name = "DrugName", dose = "DoseAsLabelled", rate =
    "InfusionRate", unit = "DoseUnitAbbreviated", route = "RouteOfMedicationAbbreviated",
    department = "DepartmentName", visit_id = "MAR_ENC_CSN", action = "MARAction")
)
}
\arguments{
\item{x}{A data frame of raw med_admin data}

\item{names}{A vector of new_name = old_name pairs}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument
}
\author{
R.J.B. Goudie
}
