% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_adt_department_col}
\alias{plot_adt_department_col}
\alias{plot_adt_department_timeline}
\title{Plot ADT department}
\usage{
plot_adt_department_col(x, facet = TRUE)

plot_adt_department_timeline(x)
}
\description{
Plot ADT department
}
