% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tests_plot_numeric}
\alias{tests_plot_numeric}
\alias{tests_histogram_time_of_day}
\alias{tests_histogram_by_date}
\alias{tests_histogram_by_weekday}
\alias{tests_plot_point}
\alias{tests_plot_tile}
\title{Tests plots}
\usage{
tests_plot_numeric(
  x,
  range_mainly_low,
  range_mainly_high,
  range_discard_below,
  range_discard_above
)

tests_histogram_time_of_day(x, datetime = collected_datetime)

tests_histogram_by_date(x, datetime = collected_datetime)

tests_histogram_by_weekday(x, datetime = collected_datetime)

tests_plot_point(x)

tests_plot_tile(x)
}
\description{
Tests plots
}
\examples{
tests <- dplyr::left_join(
  tests_magnesium_example,
  tests_info(tests_def_example))

tests_plot_numeric(tests,
                   tests$range_mainly_low,
                   tests$range_mainly_high,
                   tests$range_discard_below,
                   tests$range_discard_above)

# tests_histogram_time_of_day(tests)
# tests_histogram_by_date(tests)
# tests_histogram_by_weekday(tests)

}
