% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiology.R
\name{radiology_add}
\alias{radiology_add}
\alias{radiology_extract}
\title{Add a new definition of a radiology item}
\usage{
radiology_add(
  radiology_def,
  symbol,
  title,
  names = NA,
  search_pattern,
  search_exclude = NA,
  value_as_character_fn = narrative
)

radiology_extract(x, radiology_def, errors = stop)
}
\arguments{
\item{radiology_def}{A radiology definition}

\item{symbol}{A short name for the radiology item. This should be a
character vector that is suitable to use as an R object name}

\item{title}{The full name for the radiology item, suitable for using in
figure titles etc}

\item{names}{The names of the procedure names (\code{Proc_Name} in raw format)}

\item{search_pattern}{A charater vector of potential synonymns for this
radiology item, in case new radiology names (\code{Proc_Name} in raw format)
for the same radiology item are added in future.}

\item{search_exclude}{A character vector of radiology names
(\code{Proc_Name} in raw format) that are known NOT to be relevant.}

\item{value_as_character_fn}{A function specifying how to generate the
\code{value_as_character} column}

\item{x}{Radiology data in tidy-column-name format (after
applying \code{radiology_rename})}

\item{errors}{A function indicating what to do when an error is found}
}
\value{
A data frame with the following columns:
\code{person_id}, \code{symbol}, \code{value_as_number}, \code{value_as_character},
\code{censoring}, \code{administered_datetime},

\code{unit}, \code{name}, \code{title}, \code{method}, \code{group}, \code{type}

The result will be sorted by administered_datetime
}
\description{
Add a new definition of a radiology item

Extract radiology data into tidy format
}
\author{
R.J.B. Goudie
}
