% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiology.R
\name{radiology_unrename}
\alias{radiology_unrename}
\title{Untidy raw radiology colnames}
\usage{
radiology_unrename(x)
}
\arguments{
\item{x}{A data frame of raw radiology data with tidy names}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
Untidy raw radiology colnames
}
\author{
R.J.B. Goudie
}
