% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{default_rename}
\alias{default_rename}
\alias{adm_rename}
\alias{adt_rename}
\alias{demogs_rename}
\alias{diagnosis_pl_rename}
\alias{fsheet_io_rename}
\alias{fsheet_rename}
\alias{med_admin_rename}
\alias{med_hist_rename}
\alias{med_prescr_rename}
\alias{radiology_rename}
\alias{tests_rename}
\title{Renaming of raw column names to standardised names}
\usage{
default_rename(x)

adm_rename(x, names = default_rename("adm"))

adt_rename(x, names = default_rename("adt"))

demogs_rename(x, names = default_rename("demogs"))

diagnosis_pl_rename(x, names = default_rename("diagnosis_pl"))

fsheet_io_rename(x, names = default_rename("fsheet_io"))

fsheet_rename(x, names = default_rename("fsheet"))

med_admin_rename(x, names = default_rename("med_admin"))

med_hist_rename(x, names = default_rename("med_hist"))

med_prescr_rename(x, names = default_rename("med_prescr"))

radiology_rename(x, names = default_rename("radiology"))

tests_rename(x, names = default_rename("tests"))
}
\arguments{
\item{x}{A data frame of raw test data}

\item{names}{A vector of new_name = old_name pairs}
}
\value{
The supplied data frame, with column names in tidy-column-name format

The supplied data frame, with column names in tidy-column-name format

The supplied data frame, with column names in tidy-column-name format

The supplied data frame, with column names in tidy-column-name format

The supplied data frame, with column names in tidy-column-name format

The supplied data frame, with column names in tidy-column-name format

The supplied data frame, with column names in tidy-column-name format

The supplied data frame, with column names in tidy-column-name format

The supplied data frame, with column names in tidy-column-name format

The supplied data frame, with column names in tidy-column-name format

The supplied data frame, with column names in tidy-column-name format
}
\description{
The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument

The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument

The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument

The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument

The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument

The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument

The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument

The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument

The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument

The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument

The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument
}
\author{
R.J.B. Goudie
}
