% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\alias{default_rename}
\alias{adm_rename}
\alias{adt_rename}
\alias{demogs_rename}
\alias{diagnosis_pl_rename}
\alias{fsheet_io_rename}
\alias{fsheet_rename}
\alias{med_admin_rename}
\alias{med_hist_rename}
\alias{med_prescr_rename}
\alias{radiology_rename}
\alias{tests_rename}
\title{Rename to tidy column names}
\usage{
default_rename(x)

adm_rename(x, names = default_rename("adm"))

adt_rename(x, names = default_rename("adt"))

demogs_rename(x, names = default_rename("demogs"))

diagnosis_pl_rename(x, names = default_rename("diagnosis_pl"))

fsheet_io_rename(x, names = default_rename("fsheet_io"))

fsheet_rename(x, names = default_rename("fsheet"))

med_admin_rename(x, names = default_rename("med_admin"))

med_hist_rename(x, names = default_rename("med_hist"))

med_prescr_rename(x, names = default_rename("med_prescr"))

radiology_rename(x, names = default_rename("radiology"))

tests_rename(x, names = default_rename("tests"))
}
\arguments{
\item{x}{A data frame of raw adm data}

\item{names}{A vector of new_name = old_name pairs}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
Rename columns from the default names in the standard data format from
Clinical Informatics into standardised, shorter names (e.g. all lowercase,
underscore-separated, with \verb{_datetime} suffix etc).
}
\details{
The default renaming map can be viewed using \code{default_rename()}.

If variations from this format occur, custom renaming can be performed
using the \code{names} argument.
}
\examples{
# The raw columns names are messy: long names, mixed capitalisation, etc!
fsheet_imported_example <-
  read_tidybrookes_csv(
    file = tidybrookes_example("fsheet.csv"),
    col_types = "fsheet"
  )
colnames(fsheet_imported_example)

# Default renaming can be viewed with
default_rename("fsheet")

colnames(fsheet_rename(fsheet_imported_example))

# A warning is issued if columns are missing
fsheet_imported_example \%>\%
  select(-MEASURE_TIME) \%>\%
  fsheet_rename
}
\seealso{
\link{unrename} to reverse this process
}
\author{
R.J.B. Goudie
}
