% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{split_by_type}
\alias{split_by_type}
\title{Split a data frame into number and character value rows}
\usage{
split_by_type(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
A list with 4 components:
\itemize{
\item \code{any_numeric} a logical indicating whether any rows are numbers
\item \code{any_character} a logical indicating whether any rows are characters
\item \code{numeric} A data frame containing numeric rows
\item \code{character} A data frame containing character rows
}
}
\description{
Determine whether to use the \code{value_as_number} or \code{value_as_character}
column for each row of a data frame.
}
\details{
\itemize{
\item Where a suitable \code{type} col is present, this is used to split the rows
\item Where only a \code{value_as_number} column is present (no \code{value_as_character}
is present), then all rows are assumed numeric
\item Similar when only a \code{value_as_character} column is present
}
}
