% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{summarise_pivot_wider}
\alias{summarise_pivot_wider}
\title{Summarise (or slice) data then pivot data wider}
\usage{
summarise_pivot_wider(
  x,
  type = "summarise",
  formula,
  id_cols,
  names_from,
  values_from,
  names_suffix = NULL
)
}
\arguments{
\item{x}{A data frame}

\item{type}{Either \code{"summarise"} or \code{"slice"}}

\item{formula}{A formula specifying the summarise or slice to perform}

\item{id_cols}{A set of columns that uniquely identifies each observation.}

\item{names_from, values_from}{A pair of arguments describing which column
(or columns) to get the name of the output column (\code{group_by}), and which
column (or columns) to get the cell values from (\code{values_from}).}

\item{names_suffix}{A character string for naming the summarised output}
}
\value{
A data frame
}
\description{
Summarise (or slice if \code{type = "summarise"}) data grouped by
\code{c(id_cols, names_from)}, using the supplied \code{formula}.
}
\details{
The pivot the data wider, so that \code{names_from} appears in columns, with
the values taken from \code{values_from}

The new columns are sorted alphabetically (so that, for example, \code{x} and
\code{x_datetime} are adjacent), and column names ending in \verb{_value} are
shortened
}
\author{
R.J.B. Goudie
}
