% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{tests_extract}
\alias{tests_extract}
\title{Extract tests data into tidy format}
\usage{
tests_extract(x, tests_def, errors = stop)
}
\arguments{
\item{x}{Test data in tidy-column-name format (after applying
\code{tests_rename})}

\item{tests_def}{A tests definition}

\item{errors}{A function indicating what to do when an error is found}
}
\value{
A data frame with the following columns:
\code{person_id}, \code{symbol}, \code{value}, \code{censoring}, \code{ordered_datetime},
\code{collected_datetime}, \code{received_datetime}, \code{result_datetime},
\code{ordering_department_name}, \code{range_low}, \code{range_high}, \code{unit},
\code{name}, \code{title}, \code{method}, \code{group}
}
\description{
Extract tests data into tidy format
}
\author{
R.J.B. Goudie
}
