% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{tests_nonnumeric}
\alias{tests_nonnumeric}
\title{Helper function to extract nonnumeric test values}
\usage{
tests_nonnumeric(x)
}
\arguments{
\item{x}{A test data frame, with \code{value_as_number} column, after fixing
up various odd values etc}
}
\value{
The rows of the supplied data frame that are \code{NA} after conversion to
to numeric form
}
\description{
This tests whether everything that remains (after exclusions and conversion)
is numeric, which should be the case for \code{type = "numeric"} tests.
}
\author{
R.J.B. Goudie
}
