% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{tests_rename}
\alias{tests_rename}
\title{Tidy raw tests colnames}
\usage{
tests_rename(x, names = default_rename("tests"))
}
\arguments{
\item{x}{A data frame of raw test data}

\item{names}{A vector of new_name = old_name pairs}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument
}
\author{
R.J.B. Goudie
}
