% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{tests_unrename}
\alias{tests_unrename}
\title{Untidy raw tests colnames}
\usage{
tests_unrename(
  x,
  names = c(STUDY_SUBJECT_DIGEST = "person_id", TestName = "name", ResultValue = "value",
    ORDERED_DATETIME = "ordered_datetime", COLLECTED_DATETIME = "collected_datetime",
    RECEIVED_DATETIME = "received_datetime", ResultDate = "result_datetime",
    ORDERING_DEPARTMENT_NAME = "ordering_department", ReferenceLow = "range_low",
    ReferenceHigh = "range_high", ResultUnit = "unit", Method = "method", TestGroupName =
    "group", OrderProcId = "order_id")
)
}
\arguments{
\item{x}{A data frame of raw tests data with tidy names}

\item{names}{A vector of new_name = old_name pairs}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument
}
\author{
R.J.B. Goudie
}
