% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{unrename}
\alias{unrename}
\alias{adm_unrename}
\alias{adt_unrename}
\alias{demogs_unrename}
\alias{diagnosis_pl_unrename}
\alias{fsheet_io_unrename}
\alias{fsheet_unrename}
\alias{med_admin_unrename}
\alias{med_prescr_unrename}
\alias{radiology_unrename}
\alias{tests_unrename}
\title{Rename from tidy to raw column names}
\usage{
adm_unrename(x, names = flip_names_and_values(default_rename("adm")))

adt_unrename(x, names = flip_names_and_values(default_rename("adt")))

demogs_unrename(x, names = flip_names_and_values(default_rename("demogs")))

diagnosis_pl_unrename(
  x,
  names = flip_names_and_values(default_rename("diagnosis_pl"))
)

fsheet_io_unrename(
  x,
  names = flip_names_and_values(default_rename("fsheet_io"))
)

fsheet_unrename(x, names = flip_names_and_values(default_rename("fsheet")))

med_admin_unrename(
  x,
  names = flip_names_and_values(default_rename("med_admin"))
)

med_prescr_unrename(
  x,
  names = flip_names_and_values(default_rename("med_prescr"))
)

radiology_unrename(
  x,
  names = flip_names_and_values(default_rename("radiology"))
)

tests_unrename(
  x,
  names = c(STUDY_SUBJECT_DIGEST = "person_id", TestName = "name", ResultValue = "value",
    ORDERED_DATETIME = "ordered_datetime", COLLECTED_DATETIME = "collected_datetime",
    RECEIVED_DATETIME = "received_datetime", ResultDate = "result_datetime",
    ORDERING_DEPARTMENT_NAME = "ordering_department", ReferenceLow = "range_low",
    ReferenceHigh = "range_high", ResultUnit = "unit", Method = "method", TestGroupName =
    "group", OrderProcId = "order_id")
)
}
\arguments{
\item{x}{A data frame of raw adm data with tidy names}

\item{names}{A vector of new_name = old_name pairs}
}
\value{
The supplied data frame, with column names in tidy-column-name format
}
\description{
The standard data format from Clinical Informatics is handled by default. If
variations from this format occur, custom renaming can be performed
using the \code{names} argument
}
\examples{
fsheet_imported_example <-
  read_tidybrookes_csv(
    file = tidybrookes_example("fsheet.csv"),
    col_types = "fsheet"
  )
colnames(fsheet_imported_example)
colnames(fsheet_rename(fsheet_imported_example))
colnames(fsheet_unrename(fsheet_rename(fsheet_imported_example)))
}
\seealso{
\link{rename} to reverse this process
}
\author{
R.J.B. Goudie
}
